### 2024-07-13 b71934f
Add routines to create Eigen::tensor objects from numpy.ndarray objects; needed to set coefficient arrays from pyEXP
