### 2024-07-30 82615b1
Added RKHS EDMD methods [no ci]
