+++ pyEXP/EDMDWrappers.cc
@@ -3,6 +3,7 @@
 #include <pybind11/stl.h>
 
 #include <Koopman.H>
+#include <KoopmanRKHS.H>
 
 namespace py = pybind11;
 #include <TensorToArray.H>
@@ -292,4 +293,114 @@ void EDMDtoolkitClasses(py::module &m) {
         --------
         Koopman
         )");
+
+  py::class_<MSSA::KoopmanRKHS, std::shared_ptr<MSSA::KoopmanRKHS>> g(m, "KoopmanRKHS");
+
+  g.def(py::init<const mssaConfig&, const std::string>(),
+	R"(
+        Koopman RKHS operator approximatation class
+
+        Parameters
+        ----------
+        config : mssaConfig
+	     the input database of components
+	flag : str
+            YAML stanza of parameter values
+
+        Returns
+        -------
+        Koopman instance
+
+        Notes
+        -----
+        The configuration should be in the format:
+
+        {'example': (coefs, keylst, [])}
+
+        where keylst is a list of selected PCs.  With a SphericalBasis
+        for example, the keylst should have the format:
+
+        [[l1, m1, n1], [l2, m2, n2], ...]
+
+        Each sublist represents a PC, where l, m, and n are the 
+        spherical harmonics basis parameters.
+        )",
+	py::arg("config"),
+	py::arg("flags") = "");
+
+  g.def("eigenvalues", &KoopmanRKHS::eigenvalues,
+	R"(
+        Vector of eigenvalues from the EDMD analysis. 
+
+        Returns
+        -------
+        numpy.ndarray
+            the vector of eigenvalues
+
+        Notes
+        -----
+        Note that these eigenvalues are complex.  It may be helpful 
+        to separate the magnitude and phase using Numpy's 'absolute' 
+        and 'angle' functions.
+       )");
+
+  g.def("saveState", &KoopmanRKHS::saveState,
+	R"(
+        Save current EDMD state to an HDF5 file with the given prefix
+
+        Parameters
+        ----------
+        prefix : str
+            output filename prefix
+
+        Returns
+        -------
+        None
+        )", py::arg("prefix"));
+
+  g.def("restoreState", &KoopmanRKHS::restoreState,
+	R"(
+        Restore current EDMD state from an HDF5 file
+
+        Parameters
+        ----------
+        prefix : str
+            input filename prefix
+
+        Notes
+        -----
+	The Koopman instance must be constructed with the same data and parameters
+        as the saved state.  The restoreState routine will check for the same 
+        data dimension and trend state but can not sure	complete consistency.
+        )", py::arg("prefix"));
+
+  g.def("getModes", &KoopmanRKHS::getModes,
+	R"(
+        Access to detrended reconstructed channel series.
+
+        Returns
+        -------
+        numpy.ndarray
+            the EDMD modes
+
+        See also
+        --------
+        Use in conjunction with 'contributions' to visualize the support 
+        from each EDMD mode to the coefficient series.
+        )");
+
+  g.def("getAllKeys", &KoopmanRKHS::getAllKeys,
+	R"(
+        Provides a list of all internal channel keys (for reference)
+
+        Returns
+        -------
+        list(Key)
+            list of keys in the format described in config file
+
+        See also
+        --------
+        Koopman
+        )");
+
 }
