### 2024-08-07 caec62e
Bug fixes for KoopmanRKHS [no ci]
