+++ expui/KoopmanRKHS.cc
@@ -1,3 +1,5 @@
+#define TESTING
+
 //
 // EDMD (Koopman theory) for EXP coefficients
 //
@@ -70,26 +72,73 @@ namespace MSSA {
 
 
   //! RKHS kernel
-  double KoopmanRKHS::kernel(const Eigen::VectorXd& x, const Eigen::VectorXd& y)
+  double KoopmanRKHS::kernel(const Eigen::VectorXd& x,
+			     const Eigen::VectorXd& y)
   {
     if (rkhs == RKHS::Polynomial) {
       double prod = x.adjoint()*y;
       return pow(1.0 + prod/(d*d), alpha);
     } else if (rkhs == RKHS::Exponential) {
       double prod = x.adjoint()*y;
-      return exp(mu*prod);
+      return exp(prod/(d*d));
     } else if (rkhs == RKHS::Gaussian) {
-      auto diff = x - y;
+      Eigen::VectorXd diff = x - y;
       double prod = diff.adjoint()*diff;
-      return exp(-prod/(sigma*sigma));
+      return exp(-prod/(d*d));
     } else {
       throw std::runtime_error("KoopmanRKHS: Unknown kernel type");
     }
   }
 
+  // Structure for sorting complex numbers and retrieving a permutation index
+  using complex_elem = std::pair<std::complex<double>, int>;
+  bool complex_comparator(const complex_elem &lhs, const complex_elem &rhs)
+  {
+    auto a = lhs.first;
+    auto b = rhs.first;
+
+    double abs_a = std::abs(a);
+    double abs_b = std::abs(b);
+
+    if (abs_a == abs_b) {
+      if (std::real(a) == std::real(b))
+	return std::imag(a) < std::imag(b);
+      else
+	return std::real(a) < std::real(b);
+    } else {
+      return abs_a < abs_b;
+    }
+  }
+
+  void matrix_print(std::ostream& out, const Eigen::MatrixXcd& M)
+  {
+    for (int i=0; i<M.rows(); i++) {
+      for (int j=0; j<M.cols(); j++) {
+	out << "(" << std::setw(8) << std::fixed << std::setprecision(3) << M(i, j).real()
+	    << "," << std::setw(8) << std::fixed << std::setprecision(3) << M(i, j).imag() << ") ";
+      }
+      out << std::endl;
+    }
+  }
+
+  void matrix_idtest(std::ostream& out, const Eigen::MatrixXcd& M)
+  {
+    double max_diag = 0.0, max_offd = 0.0;
+    for (int i=0; i<M.rows(); i++) {
+      for (int j=0; j<M.cols(); j++) {
+	if (i==j) max_diag = std::max(max_diag, std::abs(M(i, j)-1.0));
+	else      max_offd = std::max(max_offd, std::abs(M(i, j)    ));
+      }
+    }
+    out << std::string(70, '-') << std::endl;
+    out << "diag diff=" << max_diag << " offd diff=" << max_offd << std::endl;
+    out << std::string(70, '-') << std::endl;
+  }
+
+
   // Algorithm 3 from Williams, Rowley, Kevrekidis
   //
-  void KoopmanRKHS::koopman_analysis()
+  void KoopmanRKHS::analysis()
   {
     // Number of channels
     //
@@ -98,35 +147,61 @@ namespace MSSA {
 
     numT = data[keys[0]].size();
 
+    // Grammian and action matrices
     G.resize(numT, numT);
     A.resize(numT, numT);
+
+    // Data matrix (rows = time, columns = keys)
     X.resize(numT, nkeys);
     
     for (int k=0; k<nkeys; k++) {
[…]
