+++ expui/KoopmanRKHS.cc
@@ -121,7 +121,7 @@ namespace MSSA {
     }
   }
 
-  void matrix_idtest(std::ostream& out, const Eigen::MatrixXcd& M)
+  bool matrix_idtest(std::ostream& out, const Eigen::MatrixXcd& M)
   {
     double max_diag = 0.0, max_offd = 0.0;
     for (int i=0; i<M.rows(); i++) {
@@ -130,9 +130,12 @@ namespace MSSA {
 	else      max_offd = std::max(max_offd, std::abs(M(i, j)    ));
       }
     }
-    out << std::string(70, '-') << std::endl;
+    out << std::string(70, '-') << std::endl << std::setprecision(6);
     out << "diag diff=" << max_diag << " offd diff=" << max_offd << std::endl;
     out << std::string(70, '-') << std::endl;
+
+    if (max_diag>1.0e-3 or max_offd>1.0e-3) return true;
+    return false;
   }
 
 
@@ -178,7 +181,35 @@ namespace MSSA {
       }
 
       if (testF) {
-	if (oscil) {
+	if (plummer) {
+	  if (radial) {
+	    for (int l=0; l<nkeys/4; l++) {
+	      double r  = data[keys[0+4*l]][i];
+	      double pr = data[keys[2+4*l]][i];
+	      double L  = data[keys[3+4*l]][i];
+
+	      Y(0+4*l) = pr;
+	      Y(1+4*l) = L/(r*r);
+	      Y(2+4*l) = -r*pow(1.0+r*r, -1.5)/tscale;
+	      Y(3+4*l) = 0.0;
+	    }
+	  } else {
+	    for (int l=0; l<nkeys/4; l++) {
+	      double x = data[keys[0+4*l]][i];
+	      double y = data[keys[1+4*l]][i];
+	      double u = data[keys[2+4*l]][i];
+	      double v = data[keys[3+4*l]][i];
+	      
+	      double r = sqrt(x*x + y*y);
+	      double dphi = -r*pow(1.0+r*r, -1.5)/tscale;
+	      
+	      Y(0+4*l) = u;
+	      Y(1+4*l) = v;
+	      Y(2+4*l) = dphi*x/r;
+	      Y(3+4*l) = dphi*y/r;
+	    }
+	  }
+	} else if (oscil) {
 	  for (int l=0; l<nkeys/4; l++) {
 	    double x0 = data[keys[0+4*l]][i];
 	    double y0 = data[keys[1+4*l]][i];
@@ -238,10 +269,10 @@ namespace MSSA {
 
     // Right and left eigenvectors
     //
-    Eigen::VectorXcd SR2 = eigensolver2.eigenvalues().conjugate();
+    Eigen::VectorXcd SR2 = eigensolver2.eigenvalues();
     Eigen::MatrixXcd V2  = eigensolver2.eigenvectors();
 
-    Eigen::VectorXcd SL2 = eigensolver3.eigenvalues();
+    Eigen::VectorXcd SL2 = eigensolver3.eigenvalues().conjugate();
     Eigen::MatrixXcd U2  = eigensolver3.eigenvectors();
 
     U  = U2;
@@ -250,7 +281,11 @@ namespace MSSA {
     SL = SL2;
     SR = SR2;
 
+    // Reorder to match eigenvalues in left and right eigenvectors
+    // using a comparison functor which sorts by magnitude, real
+    // value, and imaginary value by precidence
     {
+      // Right eigenvectors
       std::vector<complex_elem> v(SR2.size());
       for (int i=0; i<SR2.size(); i++) v[i] = complex_elem(SR2(i), i);
       std::sort(v.begin(), v.end(), complex_comparator);
@@ -259,6 +294,7 @@ namespace MSSA {
 	SR(i)    = SR2(v[i].second);
       }
 
[…]
