### 2024-08-12 031b5ff
Added Kepler test to KoopmanRKHS
