+++ expui/KoopmanRKHS.H
@@ -53,7 +53,8 @@ namespace MSSA
     //@{
     //! Test stuff
     double lam=0.9, mu=0.5, c=0.0, tscale=1.0;
-    bool plummer = true, radial = false, oscil = false, testF = true;
+    bool kepler = true, plummer = false, radial = false, oscil = false;
+    bool testF = true;
     //@}
     
 
