+++ expui/KoopmanRKHS.cc
@@ -10,7 +10,7 @@
 // trajectory matrix rather than covariance matrix analysis may lead
 // to large errors in eigenvectors.
 //
-// The implementation here follows:
+// The implementation here is based on the
 //
 // M. O. Williams, C. W. Rowley, I. G. Kevrekidis, 2015, "A
 // kernel-based method for data-driven Koopman spectral analysis",
@@ -54,8 +54,11 @@
 namespace MSSA {
 
   // Helper ostream manipulator for debug coefficient key info
+  //
   std::ostream& operator<< (std::ostream& out, const std::vector<unsigned>& t);
 
+  // Get RKHS name from enum
+  //
   std::map<KoopmanRKHS::RKHS, std::string> KoopmanRKHS::RKHS_names =
     {
       {KoopmanRKHS::RKHS::Polynomial,  "Polynomial" },
@@ -63,6 +66,8 @@ namespace MSSA {
       {KoopmanRKHS::RKHS::Gaussian,    "Gaussian"   }
     };
 
+  // Get RKHS enum from name
+  //
   std::map<std::string, KoopmanRKHS::RKHS> KoopmanRKHS::RKHS_values =
     {
       {"Polynomial",  KoopmanRKHS::RKHS::Polynomial },
@@ -71,7 +76,8 @@ namespace MSSA {
     };
 
 
-  //! RKHS kernel
+  // RKHS kernel
+  //
   double KoopmanRKHS::kernel(const Eigen::VectorXd& x,
 			     const Eigen::VectorXd& y)
   {
@@ -91,6 +97,7 @@ namespace MSSA {
   }
 
   // Structure for sorting complex numbers and retrieving a permutation index
+  //
   using complex_elem = std::pair<std::complex<double>, int>;
   bool complex_comparator(const complex_elem &lhs, const complex_elem &rhs)
   {
@@ -100,6 +107,11 @@ namespace MSSA {
     double abs_a = std::abs(a);
     double abs_b = std::abs(b);
 
+    // Three step sort:
+    // (1) By modulus
+    // (2) By real value
+    // (3) By imag value
+    //
     if (abs_a == abs_b) {
       if (std::real(a) == std::real(b))
 	return std::imag(a) < std::imag(b);
@@ -110,6 +122,8 @@ namespace MSSA {
     }
   }
 
+  // Fixed-width matrix element printing
+  //
   void matrix_print(std::ostream& out, const Eigen::MatrixXcd& M)
   {
     for (int i=0; i<M.rows(); i++) {
@@ -121,6 +135,8 @@ namespace MSSA {
     }
   }
 
+  // Check for similarity to the identity matrix
+  //
   bool matrix_idtest(std::ostream& out, const Eigen::MatrixXcd& M)
   {
     double max_diag = 0.0, max_offd = 0.0;
@@ -130,7 +146,7 @@ namespace MSSA {
 	else      max_offd = std::max(max_offd, std::abs(M(i, j)    ));
       }
     }
-    out << std::string(70, '-') << std::endl << std::setprecision(6);
+    out << std::string(70, '-') << std::endl << std::setprecision(8);
     out << "diag diff=" << max_diag << " offd diff=" << max_offd << std::endl;
     out << std::string(70, '-') << std::endl;
 
@@ -148,28 +164,36 @@ namespace MSSA {
     auto keys = getAllKeys();
     nkeys = keys.size();
 
+    // The number of time points
+    //
     numT = data[keys[0]].size();
 
-    // Grammian and action matrices
-    G.resize(numT, numT);
-    A.resize(numT, numT);
+    // Get storage for Grammian and action matrices
+    //
+    G.resize(numT, numT);	// eq. 16
[…]
