### 2024-08-24 bdaa0ab
Update to RKHS constructor and additional stdout diagnostics [no ci]
