+++ expui/LiouvilleRKHS.cc
@@ -1,6 +1,6 @@
 #define TESTING
 //
-// EDMD (Koopman theory) for EXP coefficients
+// EDMD (Liouville occupuation kernel version) for EXP coefficients
 //
 // Uses fixed-rank approximation for the SVD to save time and space.
 // Uses the approximate SVD randomized algorithm from Halko,
@@ -11,9 +11,9 @@
 //
 // The implementation here is based on the
 //
-// M. O. Williams, C. W. Rowley, I. G. Kevrekidis, 2015, "A
-// kernel-based method for data-driven Koopman spectral analysis",
-// Journal of Computational Dynamics, 2 (2), 247-265
+// Joel A. Rosenfeld and Rushikesh Kamalapurkar, "Singular Dynamic
+// Mode Decomposition", 2023, SIAM J. Appl. Dynamical Systems,
+// Vol. 22, No. 3, pp 2357-2381
 //
 
 #include <filesystem>
@@ -42,7 +42,7 @@
 
 #include <omp.h>
 
-#include <KoopmanRKHS.H>
+#include <LiouvilleRKHS.H>
 
 #include <RedSVD.H>
 #include <YamlConfig.H>
@@ -58,40 +58,43 @@ namespace MSSA {
 
   // Get RKHS name from enum
   //
-  std::map<KoopmanRKHS::RKHS, std::string> KoopmanRKHS::RKHS_names =
+  std::map<LiouvilleRKHS::RKHS, std::string> LiouvilleRKHS::RKHS_names =
     {
-      {KoopmanRKHS::RKHS::Polynomial,  "Polynomial" },
-      {KoopmanRKHS::RKHS::Exponential, "Exponential"},
-      {KoopmanRKHS::RKHS::Gaussian,    "Gaussian"   }
+      {LiouvilleRKHS::RKHS::Polynomial,  "Polynomial" },
+      {LiouvilleRKHS::RKHS::Exponential, "Exponential"},
+      {LiouvilleRKHS::RKHS::Gaussian,    "Gaussian"   }
     };
 
   // Get RKHS enum from name
   //
-  std::map<std::string, KoopmanRKHS::RKHS> KoopmanRKHS::RKHS_values =
+  std::map<std::string, LiouvilleRKHS::RKHS> LiouvilleRKHS::RKHS_values =
     {
-      {"Polynomial",  KoopmanRKHS::RKHS::Polynomial },
-      {"Exponential", KoopmanRKHS::RKHS::Exponential},
-      {"Gaussian",    KoopmanRKHS::RKHS::Gaussian   }
+      {"Polynomial",  LiouvilleRKHS::RKHS::Polynomial },
+      {"Exponential", LiouvilleRKHS::RKHS::Exponential},
+      {"Gaussian",    LiouvilleRKHS::RKHS::Gaussian   }
     };
 
 
   // RKHS kernel
   //
-  double KoopmanRKHS::kernel(const Eigen::VectorXd& x,
-			     const Eigen::VectorXd& y)
+  double LiouvilleRKHS::kernel(const Eigen::VectorXd& x,
+			       const Eigen::VectorXd& y,
+			       double mu)
   {
+    double D2 = d*d/(mu*mu);
+
     if (rkhs == RKHS::Polynomial) {
       double prod = x.adjoint()*y;
-      return pow(1.0 + prod/(d*d), alpha);
+      return pow(1.0 + prod/D2, alpha);
     } else if (rkhs == RKHS::Exponential) {
       double prod = x.adjoint()*y;
-      return exp(prod/(d*d));
+      return exp(prod/D2);
     } else if (rkhs == RKHS::Gaussian) {
       Eigen::VectorXd diff = x - y;
       double prod = diff.adjoint()*diff;
[…]
