+++ expui/LiouvilleRKHS.cc
@@ -201,9 +201,9 @@ namespace MSSA {
 	      if (t2 == 4) f2 = 2;
 	    }
 
-	    auto x1 = data[t1].row(keys[i])*rat;
-	    auto x2 = data[t2].row(keys[j])*rat;
-	    inner += kernel(data[t1].row(keys[i])*rat, data[t2].row(keys[j])*rat, mu) * f2;
+	    auto x1 = data[keys[i]].row(t1)*rat;
+	    auto x2 = data[keys[j]].row(t2)*rat;
+	    inner += kernel(x1, x2, mu) * f2;
 	  }
 
 	  G(i, j) += inner*f1;
@@ -253,8 +253,8 @@ namespace MSSA {
 	    if (t == 4) f = 2;
 	  }
 
-	  sum0 += kernel(data[0   ].row(keys[i]), data[t].row(keys[j]), mu) * f;
-	  sumT += kernel(data[nt-1].row(keys[i]), data[t].row(keys[j]), mu) * f;
+	  sum0 += kernel(data[keys[i]].row(0   ), data[keys[j]].row(t), mu) * f;
+	  sumT += kernel(data[keys[i]].row(nt-1), data[keys[j]].row(t), mu) * f;
 	}
 
 	A(i, j) += sumT - sum0;
@@ -266,6 +266,53 @@ namespace MSSA {
     double SF = (coefDB.times[1] - coefDB.times[0])/3.0;
 
     return A * SF;
+  }
+
+
+  // Compute gamma vector for RKHS space defined by mu value
+  Eigen::VectorXd LiouvilleRKHS::computeGamma
+  (const Eigen::VectorXd& x, double mu)
+  {
+    // Use Simpson's rule quadrature
+    
+    // Number of trajectories
+    //
+    auto keys = getAllKeys();
+    traj = keys.size();
+    
+    // Check for odd number of times for Simpson's 1/3 rule
+    //
+    int nt = numT;
+    if (nt/2*2 == nt) nt -= 1;
+
+    // Allocate return vector and set to zero
+    //
+    Eigen::VectorXd g(traj);
+    int f;
+
+    for (int i=0; i<traj; i++) {
+
+      g(i) = 0.0;
+
+      double sum = 0.0;
+
+      for (int t=0; t<nt; t++) {
+	
+	if (t == 0 or t == nt-1) f = 1;
+	else {
+	  if (t == 1) f = 4;
+	  if (t == 4) f = 2;
+	}
+	
+	g(i) += kernel(x, data[keys[i]].row(t), mu) * f;
+      }
+    }
+
+    // Time step factor
+    //
+    double SF = (coefDB.times[1] - coefDB.times[0])/3.0;
+
+    return g * SF;
   }
 
 
@@ -318,7 +365,9 @@ namespace MSSA {
     // Compute pseudoinvese of G2
     //
     Eigen::MatrixXd G2inv(traj, traj);
-    Eigen::Vectorxd S2inv(traj);
+    Eigen::VectorXd S2inv(traj);
+
+    double TOL = S2(S2.size()-1) * tol;
 
     for (int n=std::max<int>(0, S2.size()-evCount); n<S2.size(); n++) {
       if (S2(n) < TOL) {
@@ -331,7 +380,9 @@ namespace MSSA {
     // Compute pseudoinvese of G3
     //
     Eigen::MatrixXd G3inv(traj, traj);
-    Eigen::Vectorxd S3inv(traj);
+    Eigen::VectorXd S3inv(traj);
+
+    TOL = S3(S3.size()-1) * tol;
 
[…]
