### 2024-08-27 f2191d1
Some additional debugging; needs more explicit checking; clearly not yet passing consistency checks [no ci]
