+++ expui/LiouvilleRKHS.cc
@@ -167,17 +167,19 @@ namespace MSSA {
     // Number of trajectories
     //
     auto keys = getAllKeys();
-    traj = keys.size();
+
+    if (keys.size() != traj)
+      throw std::runtime_error("LiouvilleRKHS::computeGrammian: "
+			       "number of trajectories does not match");
     
     // Check for odd number of times for Simpson's 1/3 rule
     //
-    int nt = numT;
-    if (nt/2*2 == nt) nt -= 1;
+    int nt = numT, f1, f2;
+    if (nt % 2 == 0) nt -= 1;
 
     // Allocate Grammian matrix and set to zero
     //
     Eigen::MatrixXd G(traj, traj);
-    int f1, f2;
 
     for (int i=0; i<traj; i++) {
       for (int j=0; j<traj; j++) {
@@ -186,27 +188,23 @@ namespace MSSA {
 
 	for (int t1=0; t1<nt; t1++) {
 
-	  if (t1 == 0 or t1 == nt-1) f1 = 1;
-	  else {
-	    if (t1 == 1) f1 = 4;
-	    if (t1 == 4) f1 = 2;
-	  }
+	  if (t1 == 0 or t1 == nt-1) f1 = 1; // End points
+	  else if (t1 % 2) f1 = 2;	     // Odd interior point
+	  else f1 = 4;			     // Even interior point
 
 	  double inner = 0.0;
 	  for (int t2=0; t2<nt; t2++) {
-	    
+
 	    if (t2 == 0 or t2 == nt-1) f2 = 1;
-	    else {
-	      if (t2 == 1) f2 = 4;
-	      if (t2 == 4) f2 = 2;
-	    }
+	    else if (t2 % 2) f2 = 2;
+	    else f2 = 4;
 
 	    auto x1 = data[keys[i]].row(t1)*rat;
 	    auto x2 = data[keys[j]].row(t2)*rat;
 	    inner += kernel(x1, x2, mu) * f2;
 	  }
 
-	  G(i, j) += inner*f1;
+	  G(i, j) += inner * f1;
 	}
       }
     }
@@ -226,17 +224,18 @@ namespace MSSA {
     // Number of trajectories
     //
     auto keys = getAllKeys();
-    traj = keys.size();
+    if (traj != keys.size())
+      throw std::runtime_error("LiouvilleRKHS::computeGammaDiff: "
+			       "number of trajectories does not match");
     
     // Check for odd number of times for Simpson's 1/3 rule
     //
-    int nt = numT;
-    if (nt/2*2 == nt) nt -= 1;
+    int nt = numT, f;
+    if (nt % 2 == 0) nt -= 1;
 
     // Allocate Grammian matrix and set to zero
     //
     Eigen::MatrixXd A(traj, traj);
-    int f;
 
     for (int i=0; i<traj; i++) {
       for (int j=0; j<traj; j++) {
@@ -247,11 +246,9 @@ namespace MSSA {
 
 	for (int t=0; t<nt; t++) {
 
-	  if (t == 0 or t == nt-1) f = 1;
-	  else {
-	    if (t == 1) f = 4;
-	    if (t == 4) f = 2;
-	  }
+	  if (t == 0 or t == nt-1) f = 1; // End points
+	  else if (t % 2) f = 2;	  // Odd interior point
+	  else f = 4;			  // Even interior point
 
 	  sum0 += kernel(data[keys[i]].row(0   ), data[keys[j]].row(t), mu) * f;
 	  sumT += kernel(data[keys[i]].row(nt-1), data[keys[j]].row(t), mu) * f;
@@ -275,37 +272,73 @@ namespace MSSA {
   {
     // Use Simpson's rule quadrature
     
[…]
