### 2024-08-29 645b0f1
Added Singular Liouville method to the Liouville RKHS class [no ci]
