+++ expui/LiouvilleRKHS.cc
@@ -75,6 +75,23 @@ namespace MSSA {
     };
 
 
+  // Get Method name from enum
+  //
+  std::map<LiouvilleRKHS::Method, std::string> LiouvilleRKHS::Method_names =
+    {
+      {LiouvilleRKHS::Method::Singular,      "Singular" },
+      {LiouvilleRKHS::Method::Eigenfunction, "Eigenfunction"}
+    };
+
+  // Get RKHS enum from name
+  //
+  std::map<std::string, LiouvilleRKHS::Method> LiouvilleRKHS::Method_values =
+    {
+      {"Singular",      LiouvilleRKHS::Method::Singular   },
+      {"Eigenfunction", LiouvilleRKHS::Method::Eigenfunction}
+    };
+
+
   // RKHS kernel
   //
   double LiouvilleRKHS::kernel(const Eigen::VectorXd& x,
@@ -265,6 +282,37 @@ namespace MSSA {
     return A * SF;
   }
 
+  // Compute G matrix for RKHS space defined by mu value
+  Eigen::MatrixXd LiouvilleRKHS::computeGramGamma(double mu)
+  {
+    // Number of trajectories
+    //
+    auto keys = getAllKeys();
+    if (traj != keys.size())
+      throw std::runtime_error("LiouvilleRKHS::computeGammaDiff: "
+			       "number of trajectories does not match");
+    
+    // Allocate Grammian matrix and set to zero
+    //
+    Eigen::MatrixXd A(traj, traj);
+
+    int lT = numT - 1;
+
+    for (int i=0; i<traj; i++) {
+      for (int j=i; j<traj; j++) {
+	A(i, j) =
+	  kernel(data[keys[i]].row(lT), data[keys[j]].row(lT), mu) +
+	  kernel(data[keys[i]].row(0 ), data[keys[j]].row(0 ), mu) -
+	  kernel(data[keys[i]].row(0 ), data[keys[j]].row(lT), mu) -
+	  kernel(data[keys[i]].row(lT), data[keys[j]].row(0 ), mu) ;
+
+	if (i != j) A(j, i) = A(i, j); // Matrix is symmetric
+      }
+    }
+    
+    return A;
+  }
+
 
   // Compute gamma vector for RKHS space defined by mu value
   Eigen::VectorXd LiouvilleRKHS::computeGamma
@@ -348,10 +396,212 @@ namespace MSSA {
     return g * SF;
   }
 
+  // Compute trajectory matrix
+  Eigen::MatrixXd LiouvilleRKHS::trajectory()
+  {
+    // Number of trajectories
+    //
+    auto keys = getAllKeys();
+
+    if (traj != keys.size())
+      throw std::runtime_error("LiouvilleRKHS::trajectory: "
+			       "number of trajectories does not match");
+
+    Eigen::MatrixXd g(traj, rank);
+    g.setZero();
+    
+    int nT = numT - 1;
+
+    for (int i=0; i<traj; i++)
+      g.row(i) = data[keys[i]].row(nT) - data[keys[i]].row(0);
+
+    return g;
+  }
+
+
+  // Algorithm 6.1 from Rosenfeld and Kamalapurkhar
+  //
+  void LiouvilleRKHS::singular_analysis()
+  {
+    // Number of channels
+    //
+    auto keys = getAllKeys();
+    nkeys = keys.size();
+
+    // The number of time points
+    //
+    numT = data[keys[0]].rows();
+    rank = data[keys[0]].cols();
+
+    // Get Grammian matrices
+    //
+    G1 = computeGramGamma(mu1);
+    G2 = computeGrammian(mu2);
+
+    // For regularization
+    //
+    auto R = Eigen::MatrixXd::Identity(traj, traj) * eps;
+
[…]
