+++ expui/LiouvilleRKHS.cc
@@ -34,6 +34,8 @@
 #include <highfive/highfive.hpp>
 #include <highfive/eigen.hpp>
 
+#include <omp.h>
+
 /* For debugging
    #undef eigen_assert
    #define eigen_assert(x)						\
@@ -198,7 +200,39 @@ namespace MSSA {
     //
     Eigen::MatrixXd G(traj, traj);
 
+#pragma omp parallel for collapse(2)
     for (int i=0; i<traj; i++) {
+
+      for (int j=0; j<traj; j++) {
+
+	G(i, j) = 0.0;
+
+	for (int t1=0; t1<nt; t1++) {
+
+	  int f1 = 4;			   // Even interior point
+	  if (t1 == 0 or t1 == nt-1) f1 = 1; // End points
+	  else if (t1 % 2) f1 = 2;	   // Odd interior point
+	
+	  double inner = 0.0;
+	  for (int t2=0; t2<nt; t2++) {
+	  
+	    int f2 = 4;
+	    if (t2 == 0 or t2 == nt-1) f2 = 1;
+	    else if (t2 % 2) f2 = 2;
+	  
+	    auto x1 = data[keys[i]].row(t1)*rat;
+	    auto x2 = data[keys[j]].row(t2)*rat;
+	    inner += kernel(x1, x2, mu) * f2;
+	  }
+	
+	G(i, j) += inner * f1;
+	}
+      }
+    }
+
+    /*
+    for (int i=0; i<traj; i++) {
+
       for (int j=0; j<traj; j++) {
 
 	G(i, j) = 0.0;
@@ -208,24 +242,25 @@ namespace MSSA {
 	  if (t1 == 0 or t1 == nt-1) f1 = 1; // End points
 	  else if (t1 % 2) f1 = 2;	     // Odd interior point
 	  else f1 = 4;			     // Even interior point
-
+	  
 	  double inner = 0.0;
 	  for (int t2=0; t2<nt; t2++) {
-
+	    
 	    if (t2 == 0 or t2 == nt-1) f2 = 1;
 	    else if (t2 % 2) f2 = 2;
 	    else f2 = 4;
-
+	    
 	    auto x1 = data[keys[i]].row(t1)*rat;
 	    auto x2 = data[keys[j]].row(t2)*rat;
 	    inner += kernel(x1, x2, mu) * f2;
 	  }
-
+	  
 	  G(i, j) += inner * f1;
 	}
       }
     }
-
+    */
+      
     // Time step factor
     //
     double SF = (coefDB.times[1] - coefDB.times[0])/3.0;
@@ -254,19 +289,22 @@ namespace MSSA {
     //
     Eigen::MatrixXd A(traj, traj);
 
+#pragma omp parallel for collapse(2)
     for (int i=0; i<traj; i++) {
+
       for (int j=0; j<traj; j++) {
+
 
 	A(i, j) = 0.0;
 
 	double sumT = 0.0, sum0 = 0.0;
-
+	
 	for (int t=0; t<nt; t++) {
-
+	
+	  int  f = 4;			  // Even interior point
 	  if (t == 0 or t == nt-1) f = 1; // End points
 	  else if (t % 2) f = 2;	  // Odd interior point
-	  else f = 4;			  // Even interior point
-
+	
 	  sum0 += kernel(data[keys[i]].row(0   ), data[keys[j]].row(t), mu) * f;
 	  sumT += kernel(data[keys[i]].row(nt-1), data[keys[j]].row(t), mu) * f;
 	}
@@ -274,6 +312,30 @@ namespace MSSA {
 	A(i, j) += sumT - sum0;
       }
     }
+
+    /*
[…]
