### 2024-09-03 aeef67b
Some additional diagnostic checks [no ci]
