### 2024-09-04 191aa5d
Committing unfinished conda work
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
