+++ packaging/conda/EXP-nbody/conda_build_config.yaml
@@ -1,0 +1,10 @@
+mpi:
+  - openmpi
+  # - mpich
+
+target_platform:
+  - linux-64
+  # - linux-aarch64
+  # - linux-ppc64le
+  # - osx-64
+  # - osx-arm64
