### 2024-09-05 af414f8
Update EXP-libraries for latest change
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
