+++ pyEXP/CMakeLists.txt
@@ -1,19 +1,19 @@
 find_package(Python3 COMPONENTS Development REQUIRED)
 find_package(pybind11 REQUIRED)
 find_package(EXPutil)
-find_package(EXPcoefs)
+find_package(EXPui)
 find_package(EXPcommon)
 message(STATUS "EXPutil_FOUND=${EXPutil_FOUND} ${EXPutil_INCLUDE_DIRS} ${EXPutil_INCLUDE_DIR}")
 
 # Default build environment
 #
 # set(common_LINKLIB OpenMP::OpenMP_CXX MPI::MPI_CXX yaml-cpp
-#   EXP::EXPutil EXP::EXPcoefs EXP::EXPcommon HighFive ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES})
+#   EXP::EXPutil EXP::EXPui EXP::EXPcommon HighFive ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES})
 set(common_LINKLIB ${OpenMP_CXX_LIBRARIES} ${MPI_CXX_LIBRARIES}
   # ${YAML_CPP_LIBRARIES} # Not directly used in pyexp
   ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES}
   $<IF:$<BOOL:${EXPutil_LIBRARIES}>,${EXPutil_LIBRARIES},exputil>
-  $<IF:$<BOOL:${EXPcoefs_LIBRARIES}>,${EXPcoefs_LIBRARIES},expcoef>
+  $<IF:$<BOOL:${EXPui_LIBRARIES}>,${EXPui_LIBRARIES},expui>
   $<$<BOOL:${HighFive_FOUND}>:HighFive>)
   # $<$<BOOL:${PNG_FOUND}>:${PNG_LIBRARIES}>)
 
@@ -25,12 +25,12 @@ set(common_LINKLIB ${OpenMP_CXX_LIBRARIES} ${MPI_CXX_LIBRARIES}
 set(common_INCLUDE $<INSTALL_INTERFACE:include>
   ${DEP_INC} ${CMAKE_CURRENT_SOURCE_DIR}
   ${HDF5_INCLUDE_DIRS} ${EIGEN3_INCLUDE_DIR}
-  ${EXPutil_INCLUDE_DIRS} ${EXPcoefs_INCLUDE_DIRS} ${EXPcommon_INCLUDE_DIRS})
+  ${EXPutil_INCLUDE_DIRS} ${EXPui_INCLUDE_DIRS} ${EXPcommon_INCLUDE_DIRS})
 
 set(common_INCLUDE_DIRS_PRIVATE
     ${CMAKE_BINARY_DIR}/config)
 
-if (NOT (EXPutil_FOUND AND EXPcoefs_FOUND))
+if (NOT (EXPutil_FOUND AND EXPui_FOUND))
   list(APPEND common_INCLUDE_DIRS_PRIVATE
     $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include/>
     $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>
