+++ environment.yaml
@@ -1,0 +1,19 @@
+name: exp
+channels:
+  - conda-forge
+dependencies:
+  - python=3.12
+  - boost
+  - eigen=3
+  - fftw
+  - hdf5=1.14
+  - highfive
+  - openmpi
+  - yaml-cpp
+  - scikit-build
+  - gcc=13.2.0
+  - gfortran=13.2.0
+  - libpng
+  - libtirpc
+  - zlib
+  - gxx=13.2.0
