+++ expui/EXPcoefsConfig.cmake.in
@@ -1,13 +1,0 @@
-# Generates the cmake find_package files for EXPcoef
-
-@PACKAGE_INIT@
-
-if(NOT TARGET EXP::EXPcoefs)
-  include(@PACKAGE_CMAKEPACKAGE_INSTALL_DIR@/EXPcoefsTargets.cmake)
-endif()
-
-set_and_check(EXPcoefs_INCLUDE_DIRS @PACKAGE_CMAKE_INSTALL_PREFIX@/include/expcoefs)
-set_and_check(EXPcoefs_LIBRARY_DIRS @PACKAGE_CMAKE_INSTALL_PREFIX@/lib)
-set_and_check(EXPcoefs_LIBRARIES @PACKAGE_CMAKE_INSTALL_PREFIX@/lib/libexpcoef.so)
-
-check_required_components(EXPcoefs)

[…]
