### 2024-09-06 #6976 @dfm
:wave: @mikegrudic, @schuhmaj, @jobovy, @maxwelltsai — would any of you be willing to review this submission for The Journal of Open Source Software? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
