### 2024-09-14 56120a0
A bunch of updates for language version consistency based on a Clang 18 compile
