+++ packaging/pypi/EXP-nbody/setup.py
@@ -73,13 +73,12 @@ CopyAndSdist.file_list = [
   'LICENSE',
   'NEWS',
   'README.md',
-  # ['extern/HighFive/LICENSE', 'extern/highfive/LICENSE'],
 ]
+
 # Directories
 CopyAndSdist.directory_list = [
   'cmake',
   'src',
-  'extern/png++'
 ]
 
 # Install specifications
@@ -110,12 +109,12 @@ setup(
       "-DENABLE_PYEXP=OFF",
       "-DBUILD_UTILS=OFF",
       "-DUSE_SUBMODULES=OFF",
-      "-DENABLE_PNG=ON",
+      "-DENABLE_PNG=OFF", # png++ is only compatible with libpng 1.2.X - defunct      "-DENABLE_USER=OFF",
       "-DENABLE_USER=OFF",
       "-DENABLE_TESTS=OFF",
       "-DENABLE_XDR=ON",
       "-DENABLE_DSMC=OFF", # missing from extern so need clarification on what dependency it needs
-      "-DCMAKE_BUILD_TYPE=RELEASE"
+      "-DCMAKE_BUILD_TYPE=RELEASE",
     ]
     # extras_require={
     #   "cuda": []
