+++ src/CMakeLists.txt
@@ -1,6 +1,21 @@
 find_package(EXPutil)
 find_package(EXPui)
-find_package(EXPcommon)
+
+if(EXPutil_FOUND)
+  set(EXPUTIL EXP::EXPutil)
+  set(EXPUTIL_INCLUDE ${EXPutil_INCLUDE_DIRS})
+else()
+  set(EXPUTIL exputil) #internal name
+  set(EXPUTIL_INCLUDE $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/exputil/>)
+endif()
+
+if(EXPui_FOUND)
+  set(EXPUI EXP::EXPui)
+  set(EXPUI_INCLUDE ${EXPui_INCLUDE_DIRS})
+else()
+  set(EXPUI expui) #internal name
+  set(EXPUI_INCLUDE $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/expui/>)
+endif()
 
 set(CUDA_SRC)
 if (ENABLE_CUDA)
@@ -25,23 +40,84 @@ set(exp_SOURCES Basis.cc Bessel.cc Component.cc
   ParticleFerry.cc chkSlurm.c chkTimer.cc GravKernel.cc ${CUDA_SRC}
   CenterFile.cc PolarBasis.cc FlatDisk.cc signals.cc)
 
-set(common_INCLUDE_DIRS $<INSTALL_INTERFACE:include>
-  ${HDF5_INCLUDE_DIRS} ${EIGEN3_INCLUDE_DIR}
-  ${EXPutil_INCLUDE_DIRS} ${EXPui_INCLUDE_DIRS} ${EXPcommon_INCLUDE_DIRS})
+# set(exp_HEADERS AxisymmetricBasis.H
+# Basis.H
+# Bessel.H
+# Component.H
+# ComponentContainer.H
+# Cube.H
+# CylEXP.H
+# Cylinder.H
+# Direct.H
+# EJcom.H
+# EnumBitset.H
+# ExternalCollection.H
+# ExternalForce.H
+# FlatDisk.H
+# GravKernel.H
+# HaloBulge.H
+# MixtureBasis.H
+# NVTX.H
+# NoForce.H
+# OrbTrace.H
+# Orient.H
+# OutAscii.H
+# OutCHKPT.H
+# OutCHKPTQ.H
+# OutCalbr.H
+# OutCoef.H
+# OutDiag.H
+# OutFrac.H
+# OutLog.H
+# OutMulti.H
+# OutPS.H
+# OutPSN.H
+# OutPSP.H
+# OutPSQ.H
+# OutPSR.H
+# OutRelaxation.H
+# OutVel.H
+# Output.H
+# OutputContainer.H
+# ParticleFerry.H
+# PeriodicBC.H
+# PolarBasis.H
+# PotAccel.H
+# ScatterMFP.H
+# Shells.H
+# SlabSL.H
+# Sphere.H
+# SphericalBasis.H
+# StatsMPI.H
+# TwoCenter.H
+# TwoDCoefs.H
+# chkTimer.H
+# coef.H
+# expand.H
+# externalShock.H
+# generateRelaxation.H
+# global.H
+# global_key_set.H
+# tidalField.H)
 
-set(common_INCLUDE_DIRS_PRIVATE
-    ${CMAKE_BINARY_DIR}/config
+set(common_INCLUDE 
+  $<INSTALL_INTERFACE:include/exputil>
+  ${EXPUTIL_INCLUDE} 
+  ${EXPUI_INCLUDE}
+  ${HDF5_INCLUDE_DIRS} ${EIGEN3_INCLUDE_DIR})
+
+set(common_INCLUDE_PRIVATE
+    $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/config/>
[…]
