### 2024-09-17 f10d3a8
Fix pyEXP pypi packaging
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
