+++ pyEXP/CMakeLists.txt
@@ -1,78 +1,69 @@
-find_package(Python3 COMPONENTS Development REQUIRED)
+find_package(Python3 REQUIRED)
+find_package(PythonExtensions)
+
 
 if (USE_SUBMODULES)
   message("Configuring pybind11 for ${GIT_BRANCH}/${GIT_COMMIT} at ${COMPILE_TIME}")
   include_directories(${PROJECT_SOURCE_DIR}/extern/pybind11/include)
   add_subdirectory(extern/pybind11)
 else()
-  find_package(pybind11 REQUIRED)
+  find_package(pybind11 CONFIG REQUIRED)
 endif()
-find_package(EXPutil)
-find_package(EXPui)
-find_package(EXPcommon)
-message(STATUS "EXPutil_FOUND=${EXPutil_FOUND} ${EXPutil_INCLUDE_DIRS} ${EXPutil_INCLUDE_DIR}")
 
 # Default build environment
-#
-# set(common_LINKLIB OpenMP::OpenMP_CXX MPI::MPI_CXX yaml-cpp
-#   EXP::EXPutil EXP::EXPui EXP::EXPcommon HighFive ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES})
-set(common_LINKLIB ${OpenMP_CXX_LIBRARIES} ${MPI_CXX_LIBRARIES}
-  # ${YAML_CPP_LIBRARIES} # Not directly used in pyexp
+set(common_LINKLIBS ${OpenMP_CXX_LIBRARIES} ${MPI_CXX_LIBRARIES}
   ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES}
-  $<IF:$<BOOL:${EXPutil_LIBRARIES}>,${EXPutil_LIBRARIES},exputil>
-  $<IF:$<BOOL:${EXPui_LIBRARIES}>,${EXPui_LIBRARIES},expui>
   $<$<BOOL:${HighFive_FOUND}>:HighFive>)
-  # $<$<BOOL:${PNG_FOUND}>:${PNG_LIBRARIES}>)
-
-
-# if(PNG_FOUND)
-#   list(APPEND common_LINKLIB PNG::PNG)
-# endif()
 
 set(common_INCLUDE $<INSTALL_INTERFACE:include>
   ${DEP_INC} ${CMAKE_CURRENT_SOURCE_DIR}
   ${HDF5_INCLUDE_DIRS} ${EIGEN3_INCLUDE_DIR}
   ${EXPutil_INCLUDE_DIRS} ${EXPui_INCLUDE_DIRS} ${EXPcommon_INCLUDE_DIRS})
 
-set(common_INCLUDE_DIRS_PRIVATE
+set(common_INCLUDE_PRIVATE
     ${CMAKE_BINARY_DIR}/config)
 
-if (NOT (EXPutil_FOUND AND EXPui_FOUND))
-  list(APPEND common_INCLUDE_DIRS_PRIVATE
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include/>
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}>
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
+find_package(EXPutil)
+if(EXPutil_FOUND)
+  list(APPEND common_LINKLIBS EXP::EXPutil)
+  list(APPEND common_INCLUDE ${EXPutil_INCLUDE_DIRS})
+else()
+  list(APPEND common_LINKLIBS exputil) #internal name
[…]
