### 2024-09-17 158a7ff
Fix pybind11 load when using submodules
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
