+++ pyEXP/CMakeLists.txt
@@ -4,8 +4,8 @@ find_package(PythonExtensions)
 
 if (USE_SUBMODULES)
   message("Configuring pybind11 for ${GIT_BRANCH}/${GIT_COMMIT} at ${COMPILE_TIME}")
+  add_subdirectory(${PROJECT_SOURCE_DIR}/extern/pybind11)
   include_directories(${PROJECT_SOURCE_DIR}/extern/pybind11/include)
-  add_subdirectory(extern/pybind11)
 else()
   find_package(pybind11 CONFIG REQUIRED)
 endif()
