### 2024-09-17 780f36a
Silence find_package warnings when not needed
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
