### 2024-09-17 a3d6c9b
Move pybind11 subdir add to outer CMake file
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
