+++ exputil/CMakeLists.txt
@@ -52,6 +52,12 @@ set(common_INCLUDE_PRIVATE
   $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/extern/>
   $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include/>)
 
+if(USE_SUBMODULES)
+  list(APPEND common_INCLUDE_PRIVATE 
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/yaml-cpp/include>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>)
+endif()
+
 set(common_LINKLIBS
   ${DEP_LIB} ${OpenMP_CXX_LIBRARIES} ${MPI_CXX_LIBRARIES}
   ${YAML_CPP_LIBRARIES} ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES}
