### 2024-09-17 c0d294a
Add config dir for whole project
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
