### 2024-09-17 36f191c
Use build interface generator function for config include
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
