### 2024-09-18 60d7fb4
Adjust fortran and libtirpc strategy
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
