+++ packaging/pypi/pyEXP/setup.py
@@ -91,6 +91,9 @@ setup(
     license="GPL-3.0",
     packages=["pyEXP"],
     python_requires=">=3.8",
+    install_requires=[
+      'exp-libraries'
+    ],
     package_dir={
       "pyEXP": "pyEXP"
     },
@@ -114,6 +117,7 @@ setup(
       "-DENABLE_TESTS=OFF",
       "-DENABLE_XDR=ON",
       "-DENABLE_DSMC=OFF", # missing from extern so need clarification on what dependency it needs
-      "-DCMAKE_BUILD_TYPE=RELEASE"
+      "-DCMAKE_BUILD_TYPE=RELEASE",
+      "-Wno-dev"
     ]
 )
