+++ exputil/CMakeLists.txt
@@ -52,17 +52,18 @@ set(common_INCLUDE_PRIVATE
   $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/extern/>
   $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include/>)
 
-if(USE_SUBMODULES)
-  list(APPEND common_INCLUDE_PRIVATE 
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/yaml-cpp/include>
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>)
-endif()
-
 set(common_LINKLIBS
   ${DEP_LIB} ${OpenMP_CXX_LIBRARIES} ${MPI_CXX_LIBRARIES}
   ${YAML_CPP_LIBRARIES} ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES}
   ${FFTW_DOUBLE_LIB} 
   $<$<BOOL:${HighFive_FOUND}>:HighFive>)
+
+if(USE_SUBMODULES)
+  list(APPEND common_INCLUDE_PRIVATE 
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/yaml-cpp/include>
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>)
+  list(APPEND common_LINKLIBS yaml-cpp)
+endif()
 
 if(ENABLE_CUDA)
   list(APPEND common_INCLUDE ${CMAKE_CUDA_TOOLKIT_INCLUDE_DIRECTORIES} ${CUDAToolkit_INCLUDE_DIRS})
