+++ expui/EXPuiConfig.cmake.in
@@ -11,3 +11,7 @@ set_and_check(EXPui_LIBRARY_DIRS @PACKAGE_CMAKE_INSTALL_PREFIX@/lib)
 set_and_check(EXPui_LIBRARIES @PACKAGE_CMAKE_INSTALL_PREFIX@/lib/libexpui.so)
 
 check_required_components(EXPui)
+
+if (EXPui_FOUND)
+    find_package_message(EXPui "Found EXPui: ${EXPui_LIBRARIES}" "[${EXPui_LIBRARIES}][${EXPui_INCLUDE_DIRS}]")
+endif()

[…]
