+++ utils/ICs/CMakeLists.txt
@@ -19,7 +19,7 @@ endif()
 set(common_INCLUDE
   $<INSTALL_INTERFACE:include>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/DSMC/src>
-  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>)
+  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/utils/SL/>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include/>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/>
