+++ utils/SL/CMakeLists.txt
@@ -16,6 +16,7 @@ endif()
 set(common_INCLUDE
   $<INSTALL_INTERFACE:include>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/DSMC/>
+  $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include/>
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/>
   ${CMAKE_BINARY_DIR} ${DEP_INC}
