### 2024-09-23 479e1c1
Update tests for new pyEXP layout
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
