+++ expui/CMakeLists.txt
@@ -12,7 +12,7 @@ set(common_LINKLIBS ${DEP_LIB} ${OpenMP_CXX_LIBRARIES}
   ${MPI_CXX_LIBRARIES}
   ${YAML_CPP_LIBRARIES} ${EXPUTIL}
   ${VTK_LIBRARIES} ${HDF5_LIBRARIES} ${HDF5_HL_LIBRARIES}
-  $<$<BOOL:${HighFive_FOUND}>:HighFive>)
+  HighFive)
 
 set(common_INCLUDE 
   ${DEP_INC} ${HDF5_INCLUDE_DIRS}
@@ -23,10 +23,15 @@ set(common_INCLUDE_PRIVATE
   $<BUILD_INTERFACE:${CMAKE_BINARY_DIR}/config/>
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
 
+#   if(HighFive_FOUND)
+#   list(APPEND common_LINKLIBS HighFive)
+# else()
+#   list(APPEND common_INCLUDE_PRIVATE $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>)
+# endif()
+
 if(USE_SUBMODULES)
   list(APPEND common_INCLUDE_PRIVATE 
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/yaml-cpp/include>
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/HighFive/include>)
+    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/extern/yaml-cpp/include>)
   list(APPEND common_LINKLIBS yaml-cpp)
 endif()
 
