### 2024-09-26 9739781
Use Eigen::Ref<> to pass arrays to prevent pybind11 mapping failure in pyEXP
