+++ expui/Coefficients.cc
@@ -469,7 +469,7 @@ namespace CoefClasses
     return mat;
   }
   
-  void SphCoefs::setData(double time, const Eigen::VectorXcd& dat)
+  void SphCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -484,7 +484,7 @@ namespace CoefClasses
     }
   }
   
-  void SphCoefs::setMatrix(double time, const Eigen::MatrixXcd& dat)
+  void SphCoefs::setMatrix(double time, Eigen::Ref<Eigen::MatrixXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -891,7 +891,7 @@ namespace CoefClasses
     return mat;
   }
   
-  void CylCoefs::setData(double time, const Eigen::VectorXcd& dat)
+  void CylCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -906,7 +906,7 @@ namespace CoefClasses
     }
   }
 
-  void CylCoefs::setMatrix(double time, const Eigen::MatrixXcd& dat)
+  void CylCoefs::setMatrix(double time, Eigen::Ref<Eigen::MatrixXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1247,7 +1247,7 @@ namespace CoefClasses
     return arr;
   }
 
-  void SlabCoefs::setData(double time, const Eigen::VectorXcd& dat)
+  void SlabCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1599,7 +1599,7 @@ namespace CoefClasses
     return arr;
   }
 
-  void CubeCoefs::setData(double time, const Eigen::VectorXcd& dat)
+  void CubeCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1968,7 +1968,7 @@ namespace CoefClasses
     return arr;
   }
   
-  void TableData::setData(double time, const Eigen::VectorXcd& dat)
+  void TableData::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -2458,7 +2458,7 @@ namespace CoefClasses
     return *mat;
   }
   
-  void SphFldCoefs::setData(double time, const Eigen::VectorXcd& dat)
+  void SphFldCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -2897,7 +2897,7 @@ namespace CoefClasses
     return *mat;
   }
   
-  void CylFldCoefs::setData(double time, const Eigen::VectorXcd& dat)
+  void CylFldCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
   {
     auto it = coefs.find(roundTime(time));
 
