+++ expui/Coefficients.H
@@ -111,7 +111,7 @@ namespace CoefClasses
     virtual Eigen::VectorXcd& getData(double time) = 0;
 
     //! Set coefficient store at given time to the provided matrix
-    virtual void setData(double time, const Eigen::VectorXcd& data) = 0;
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> data) = 0;
 
     //! Interpolate coefficient matrix at given time
     std::tuple<Eigen::VectorXcd&, bool> interpolate(double time);
@@ -273,10 +273,10 @@ namespace CoefClasses
     Eigen::MatrixXcd& getMatrix(double time);
 
     //! Set coefficient matrix at given time
-    virtual void setData(double time, const Eigen::VectorXcd& arr);
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> arr);
 
     //! Natural data setter for pyEXP
-    void setMatrix(double time, const Eigen::MatrixXcd& mat);
+    void setMatrix(double time, Eigen::Ref<Eigen::MatrixXcd> mat);
 
     //! Interpolate coefficient matrix at given time
     std::tuple<Eigen::MatrixXcd&, bool> interpolate(double time)
@@ -411,10 +411,10 @@ namespace CoefClasses
     Eigen::MatrixXcd& getMatrix(double time);
     
     //! Set coefficient matrix at given time
-    virtual void setData(double time, const Eigen::VectorXcd& arr);
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> arr);
     
     //! For pyEXP
-    void setMatrix(double time, const Eigen::MatrixXcd& arr);
+    void setMatrix(double time, Eigen::Ref<Eigen::MatrixXcd> arr);
     
     //! Interpolate coefficient matrix at given time
     std::tuple<Eigen::MatrixXcd&, bool> interpolate(double time)
@@ -542,7 +542,7 @@ namespace CoefClasses
     virtual Eigen3d& getTensor(double time);
     
     //! Set coefficient matrix at given time
-    virtual void setData(double time, const Eigen::VectorXcd& dat);
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> dat);
 
     //! Native version
     virtual void setTensor(double time, const Eigen3d& dat);
@@ -673,7 +673,7 @@ namespace CoefClasses
     virtual Eigen3d& getTensor(double time);
     
     //! Set coefficient matrix at given time
-    virtual void setData(double time, const Eigen::VectorXcd& dat);
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> dat);
 
     //! Native version
     virtual void setTensor(double time, const Eigen3d& dat);
@@ -810,9 +810,7 @@ namespace CoefClasses
     /** Set coefficient matrix at given time.  This is for pybind11,
 	since the operator() will not allow lvalue assignment, it
 	seems. */
-    virtual void setData(double time, const Eigen::VectorXcd& arr);
-
-    void setArray(double time, const Eigen::VectorXcd& arr) { setData(time, arr); }
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> arr);
 
     //! Get coefficient structure at a given time
     virtual std::shared_ptr<CoefStruct> getCoefStruct(double time)
@@ -914,7 +912,7 @@ namespace CoefClasses
     SphFldStruct::coefType& getMatrix(double time);
 
     //! Set coefficient matrix at given time
-    virtual void setData(double time, const Eigen::VectorXcd& arr);
+    virtual void setData(double time, Eigen::Ref<Eigen::VectorXcd> arr);
 
     //! Natural data setter for pyEXP
[…]
