### 2024-09-26 0b04688
Use Eigen::Ref<> to pass arrays to prevent pybind11 mapping failure in pyEXP
