### 2024-09-26 d583a10
Fix norm issue discovered in C++ tests
