+++ exputil/EmpCyl2d.cc
@@ -638,7 +638,7 @@ std::shared_ptr<EmpCyl2d::ModelCyl> EmpCyl2d::createModel()
   if (name.find("kuzmin") != std::string::npos) {
     if (myid==0) {
       std::cout << "---- EmpCyl2d::ModelCyl: Making a Kuzmin disk";
-      if (Params["parameters"]) std::cout << " with " << Params["parameters"];
+      if (Params["parameters"]) std::cout << " with" << std::endl << Params["parameters"];
       std::cout << std::endl;
     }
     return std::make_shared<KuzminCyl>(Params["parameters"]);
@@ -647,7 +647,7 @@ std::shared_ptr<EmpCyl2d::ModelCyl> EmpCyl2d::createModel()
   if (name.find("mestel") != std::string::npos) {
     if (myid==0) {
       std::cout << "---- EmpCyl2d::ModelCyl: Making a finite Mestel disk";
-      if (Params["parameters"]) std::cout << " with " << Params["parameters"];
+      if (Params["parameters"]) std::cout << " with" << std::endl << Params["parameters"];
       std::cout << std::endl;
     }
     return std::make_shared<MestelCyl>(Params["parameters"]);
@@ -656,7 +656,9 @@ std::shared_ptr<EmpCyl2d::ModelCyl> EmpCyl2d::createModel()
   if (name.find("zang") != std::string::npos) {
     if (myid==0) {
       std::cout << "---- EmpCyl2d::ModelCyl: Making a double-tapered Zang";
-      if (Params["parameters"]) std::cout << " with " << Params["parameters"];
+      YAML::Emitter out;
+      out << YAML::Flow << Params["parameters"];
+      if (Params["parameters"]) std::cout << " with " << out.c_str();
       std::cout << std::endl;
     }
     return std::make_shared<ZangCyl>(Params["parameters"]);
@@ -665,7 +667,7 @@ std::shared_ptr<EmpCyl2d::ModelCyl> EmpCyl2d::createModel()
   if (name.find("expon") != std::string::npos) {
     if (myid==0) {
       std::cout << "---- EmpCyl2d::ModelCyl: Making an Exponential disk";
-      if (Params["parameters"]) std::cout << " with " << Params["parameters"];
+      if (Params["parameters"]) std::cout << " with" << std::endl << Params["parameters"];
       std::cout << std::endl;
     }
     return std::make_shared<ExponCyl>(Params["parameters"]);
@@ -834,7 +836,7 @@ void EmpCyl2d::create_tables()
 	for (int l=0; l<nmaxfid; l++) {
 	  D(j, l) += fac * disk->dens(rr) *
 	    basis->potl(m, j, rr) * basis->potl(m, l, rr)
-	    / sqrt(basis->norm(j, m)*basis->norm(l, m));
+	    / sqrt(basis->norm(j, m)*basis->norm(l, m)*4.0);
 	}
       }
     }
@@ -871,9 +873,9 @@ void EmpCyl2d::create_tables()
       if (m==0) xgrid[i] = r;
 
       for (int n=0; n<nmaxfid; n++) {
-	pot(n) = basis->potl(m, n, r) / sqrt(basis->norm(n, m));
-	den(n) = basis->dens(m, n, r) / sqrt(basis->norm(n, m));
-	dph(n) = basis->dpot(m, n, r) / sqrt(basis->norm(n, m));
[…]
