### 2024-09-26 0fea8e3
Fixes for H5 API
