### 2024-09-26 6dad8ce
Do not copy data if nimax or ndmax are zero
