### 2024-09-27 b30361d
OutVel should put its file in outdir for consistence
