### 2024-09-28 876edb6
Remove Eigen::Ref from setMatrix to allow more flexibility
