### 2024-09-28 8ab5ba2
Fix typos in diagnostic errors [no ci]
