### 2024-09-28 12176dc
Change names of Eigen::Tensor-->ndarray generators for clarity
