+++ expui/Coefficients.cc
@@ -469,13 +469,13 @@ namespace CoefClasses
     return mat;
   }
   
-  void SphCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
+  void SphCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
     if (it == coefs.end()) {
       std::ostringstream str;
-      str << "SphCoefs::setMatrix: requested time=" << time << " not found";
+      str << "SphCoefs::setData: requested time=" << time << " not found";
       throw std::runtime_error(str.str());
     } else {
       it->second->store = dat;
@@ -484,7 +484,7 @@ namespace CoefClasses
     }
   }
   
-  void SphCoefs::setMatrix(double time, Eigen::Ref<Eigen::MatrixXcd> dat)
+  void SphCoefs::setMatrix(double time, Eigen::MatrixXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -891,13 +891,13 @@ namespace CoefClasses
     return mat;
   }
   
-  void CylCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
+  void CylCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
     if (it == coefs.end()) {
       std::ostringstream str;
-      str << "CylCoefs::setMatrix: requested time=" << time << " not found";
+      str << "CylCoefs::setData: requested time=" << time << " not found";
       throw std::runtime_error(str.str());
     } else {
       it->second->store = dat;
@@ -906,7 +906,7 @@ namespace CoefClasses
     }
   }
 
-  void CylCoefs::setMatrix(double time, Eigen::Ref<Eigen::MatrixXcd> dat)
+  void CylCoefs::setMatrix(double time, Eigen::MatrixXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1247,13 +1247,13 @@ namespace CoefClasses
     return arr;
   }
 
-  void SlabCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
+  void SlabCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
     if (it == coefs.end()) {
       std::ostringstream str;
-      str << "CylCoefs::setMatrix: requested time=" << time << " not found";
+      str << "CylCoefs::setData: requested time=" << time << " not found";
       throw std::runtime_error(str.str());
     } else {
       it->second->store = dat;
@@ -1268,7 +1268,7 @@ namespace CoefClasses
 
     if (it == coefs.end()) {
       std::ostringstream str;
-      str << "CylCoefs::setMatrix: requested time=" << time << " not found";
+      str << "SlabCoefs::setTensor: requested time=" << time << " not found";
       throw std::runtime_error(str.str());
     } else {
       it->second->allocate();	// Assign storage for the flattened tensor
@@ -1599,13 +1599,13 @@ namespace CoefClasses
     return arr;
   }
 
-  void CubeCoefs::setData(double time, Eigen::Ref<Eigen::VectorXcd> dat)
+  void CubeCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
     if (it == coefs.end()) {
       std::ostringstream str;
-      str << "CylCoefs::setMatrix: requested time=" << time << " not found";
+      str << "CubeCoefs::setData: requested time=" << time << " not found";
       throw std::runtime_error(str.str());
     } else {
       it->second->store = dat;
[…]
