### 2024-09-29 #7302 @editorialbot
[REVIEW]: EXP: a Python/C++ package for basis function expansion methods in galactic dynamics
**Submitting author:** @michael-petersen (<a href="http://orcid.org/0000-0003-1517-3935">Michael Petersen</a>)
**Repository:** https://github.com/EXP-code/EXP
**Branch with paper.md** (empty if default branch):
**Version:** 7.8.0
