### 2024-10-05 828f583
Change to variable number of fields in SphFldStruct and CylFldStruct constructors
