### 2024-10-05 1212151
Change names of Eigen::Tensor wrappers for consistency
