### 2024-10-06 c59e83f
Add spack package draft
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
