### 2024-10-08 7aafdfd
Added test Dockerfile for binder [no ci]
