+++ Dockerfile
@@ -1,0 +1,168 @@
+# Image for building EXP with MPI and CUDA based on the hpcbase recipe
+#    and our standard Ubuntu 22.04 development environment for Docker
+# 
+# Contents:
+# 
+#   * Ubuntu 22.04 
+#   * GNU compilers (upstream)
+#   * Python 2 and 3 (upstream)
+#   * OpenMPI, FFTW3, HDF5, Eigen3, PNG (upstream)
+# 
+# Build notes:
+# 
+#   * The container build requires the following commands:
+#     $ hpccm --recipe exp_all_deb.py --format docker > Dockerfile
+#     $ docker build -t exp-test -f Dockerfile .
+#   or
+#     $ docker buildx build --platform=linux/amd64,linux/arm64 -t the9cat/exp:latest --push -f Dockerfile .
+# 
+#   * You will need to put the EXP.tar.gz file in the build directory. I
+#     like to make a fresh clone and run `git submodule update --init
+#     --recursive` before tarring it up.
+# 
+# 
+
+FROM ubuntu:22.04 AS devel
+
+# Python
+RUN apt-get update -y && \
+    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+        python2 \
+        python3 && \
+    rm -rf /var/lib/apt/lists/*
+
+# GNU compiler
+RUN apt-get update -y && \
+    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+        g++ \
+        gcc \
+        gfortran && \
+    rm -rf /var/lib/apt/lists/*
+
+RUN apt-get update -y && \
+    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+        cmake \
+        git \
+        libeigen3-dev \
+        libfftw3-dev \
+        libhdf5-103 \
+        libhdf5-dev \
+        libopenmpi-dev \
+        libpng-dev \
+        make \
+        openmpi-bin \
+        python3-dev \
+        tar \
+        wget && \
+    rm -rf /var/lib/apt/lists/*
+
+# EXP.tar.gz
+COPY EXP.tar.gz /var/tmp/EXP.tar.gz
+RUN mkdir -p /var/tmp && tar -x -f /var/tmp/EXP.tar.gz -C /var/tmp -z && \
+    cd /var/tmp/EXP && \
+    git config --global --add safe.directory /var/tmp/EXP && \
+    git config --global --add safe.directory /var/tmp/EXP/extern/HighFive && \
+    git config --global --add safe.directory /var/tmp/EXP/extern/pybind11 && \
+    git config --global --add safe.directory /var/tmp/EXP/extern/yaml-cpp && \
+    git config --global --add safe.directory /var/tmp/EXP/extern/HighFive/deps/catch2 && \
+    mkdir -p /usr/local/EXP/doc && \
+    cp -a /var/tmp/EXP/sphinx/* /usr/local/EXP/doc && \
+    mkdir -p /var/tmp/EXP/build && cd /var/tmp/EXP/build && cmake -DCMAKE_INSTALL_PREFIX=/usr/local/EXP -D CMAKE_BUILD_TYPE=Release -D ENABLE_CUDA=NO -D ENABLE_USER=YES -D ENABLE_SLURM=NO -D ENABLE_PNG=NO -D ENABLE_VTK=NO -D FFTW_INCLUDE_DIRS=/usr/include/fftw3 -D Eigen3_DIR=/usr/share/eigen3/cmake -D CMAKE_INSTALL_PREFIX=/usr/local/EXP /var/tmp/EXP && \
[…]
