### 2024-10-08 6e3d628
Use git clone and disable SSL [no ci]
