### 2024-10-09 db9d0ee
Added EXP-examples and pyEXP-examples to Docker image for binder [no ci]
