+++ include/BiorthCyl.H
@@ -5,6 +5,7 @@
 #include <fstream>
 #include <string>
 #include <cmath>
+#include <any>
 
 #include <Eigen/Eigen>
 
@@ -173,8 +174,7 @@ public:
   cacheInfo(const std::string& cachefile, bool verbose=true);
 
 #if HAVE_LIBCUDA==1
-  void initialize_cuda(std::vector<cudaArray_t>& cuArray,
-		       thrust::host_vector<cudaTextureObject_t>& tex);
+  void initialize_cuda(std::vector<cudaArray_t>& cuArray, std::any tex);
 
   virtual cudaMappingConstants getCudaMappingConstants()
   {

[…]
