### 2024-10-10 d006e86
Implement that CUDA 16 change in NVTX for nvtx3; run CUDA-aware source code through nvcc for consistence with the thrust and cub ABI changes
