+++ src/ExternalForce.H
@@ -53,6 +53,7 @@ public:
   //! Finish and clean-up (caching data necessary for restart)
   virtual void finish() {}
 
+  // #if HAVE_LIBCUDA==1 && defined (__NVCC__)
 #if HAVE_LIBCUDA==1
   //! Copy particles from device for non-cuda forces
   void getParticlesCuda(Component *c);
