### 2024-10-10 cdd2897
Do not enable CUDA twice
