### 2024-10-10 d24ae4f
Put Sphere.cc in nvcc compile list to prevent issues with parameter passing
