### 2024-10-10 b0f0f6e
Beginning to implement that std::any strategy
