### 2024-10-12 14a6ef5
Added CTests for 'setMatrix()' and 'createFromArray' [no ci]
