### 2024-10-12 ae8a90b
Update build.yml
Show the output in the log for failing ctests
