### 2024-10-12 88d1338
Make ctest log the output on failure
