### 2024-10-12 41db618
Remove intentional failure from test script
