### 2024-10-13 e9600e4
Exit after basis creation; sanity check
