### 2024-10-13 eaa6521
Exit after data creation; sanity check
