+++ tests/Halo/createCoefs.py
@@ -27,8 +27,6 @@ basis = pyEXP.basis.Basis.factory(config)
 
 print("---- created basis")
 
-exit(0)
-
 # Make some fake body data
 #
 
@@ -45,6 +43,8 @@ for i in range(0, 100):
     xpos.append(random.random()*2.0 - 1.0)
     ypos.append(random.random()*2.0 - 1.0)
     zpos.append(random.random()*2.0 - 1.0)
+
+exit(0)
 
 print("---- createFromArray usings lists")
 coef1 = basis.createFromArray(mass, [xpos, ypos, zpos], time=3.0)
