### 2024-10-13 6432520
Exit after first coef creation; sanity check
