### 2024-10-13 8d332a2
Exit after third coef creation; sanity check
