### 2024-10-13 2369691
Exit after second coef creation; sanity check
