### 2024-10-13 747e6a5
Comment out second coef creation; sanity check
