### 2024-10-13 1043436
Comment out all but first coef creation; sanity check
