### 2024-10-13 278e1fe
Restore original test script
