### 2024-10-13 4c27205
Update createCoefs.py
One more try just to make sure that this fails on all `createFromArrry` statements.
