### 2024-10-13 6853072
Update createCoefs.py
More bisection...
