+++ tests/Halo/createCoefs.py
@@ -47,7 +47,11 @@ for i in range(0, 100):
 print("---- createFromArray usings lists")
 coef1 = basis.createFromArray(mass, [xpos, ypos, zpos], time=3.0)
 
-exit(0) # TEST
+coefs = pyEXP.coefs.SphCoefs(True)
+coefs.add(coef1)
+
+print("Times:", coefs.Times())
+exit(0) # TEST END
 
 mass  = np.array(mass)
 data  = np.array([xpos, ypos, zpos])
