### 2024-10-13 ca80b09
Update createCoefs.py
Another bisection, renaming arrays.
