### 2024-10-13 7035069
Update createCoefs.py
More updates and bisection...
