+++ tests/Halo/createCoefs.py
@@ -58,29 +58,30 @@ coef2 = basis.createFromArray(mass2, data2, time=3.1)
 
 coefs.add(coef2)
 
-print("Times:", coefs.Times())
-exit(0) # TEST END
-
-mass = np.ones(1000) * 1.0e6
-xpos = np.random.normal(0.0, 1.0, 1000)
-ypos = np.random.normal(0.0, 1.0, 1000)
-zpos = np.random.normal(0.0, 1.0, 1000)
+mass3 = np.ones(1000) * 1.0e6
+xpos3 = np.random.normal(0.0, 1.0, 1000)
+ypos3 = np.random.normal(0.0, 1.0, 1000)
+zpos3 = np.random.normal(0.0, 1.0, 1000)
 
 print("---- createFromArray using a list of numpy arrays")
-coef3 = basis.createFromArray(mass, [xpos, ypos, zpos], time=3.2)
+coef3 = basis.createFromArray(mass3, [xpos3, ypos3, zpos3], time=3.2)
 
-data  = np.array([xpos, ypos, zpos])
+coefs.add(coef3)
+
+data4  = np.array([xpos3, ypos3, zpos3])
 
 print("---- createFromArray usings pure numpy arrays")
-coef4 = basis.createFromArray(mass, data, time=3.3)
+coef4 = basis.createFromArray(mass3, data4, time=3.3)
+
+coefs.add(coef4)
 
 # Add the coefficient structure coefficient structure
 #
-coefs = pyEXP.coefs.SphCoefs(True)
-coefs.add(coef1)
-coefs.add(coef2)
-coefs.add(coef3)
-coefs.add(coef4)
+# coefs = pyEXP.coefs.SphCoefs(True)
+# coefs.add(coef1)
+# coefs.add(coef2)
+# coefs.add(coef3)
+# coefs.add(coef4)
 print("Times:", coefs.Times())
 
 exit(0)
