+++ tests/Halo/createCoefs.py
@@ -27,8 +27,11 @@ basis = pyEXP.basis.Basis.factory(config)
 
 print("---- created basis")
 
-# Make some fake body data
+# Create a coefficient structure
 #
+coefs = pyEXP.coefs.SphCoefs(True)
+
+print("---- created coefficients")
 
 # Call the basis to generate coefficients
 #
@@ -47,8 +50,9 @@ for i in range(0, 100):
 print("---- createFromArray usings lists")
 coef1 = basis.createFromArray(mass1, [xpos1, ypos1, zpos1], time=3.0)
 
-coefs = pyEXP.coefs.SphCoefs(True)
 coefs.add(coef1)
+
+print("---- creating array data from list data")
 
 mass2 = np.array(mass1)
 data2 = np.array([xpos1, ypos1, zpos1])
@@ -57,6 +61,8 @@ print("---- createFromArray usings numpy arrays converted from lists")
 coef2 = basis.createFromArray(mass2, data2, time=3.1)
 
 coefs.add(coef2)
+
+print("---- creating array data from list of numpy arrays")
 
 mass3 = np.ones(1000) * 1.0e6
 xpos3 = np.random.normal(0.0, 1.0, 1000)
@@ -69,19 +75,15 @@ coef3 = basis.createFromArray(mass3, [xpos3, ypos3, zpos3], time=3.2)
 coefs.add(coef3)
 
 data4  = np.array([xpos3, ypos3, zpos3])
+print("---- position data shape is:", data4.shape)
 
 print("---- createFromArray usings pure numpy arrays")
 coef4 = basis.createFromArray(mass3, data4, time=3.3)
 
 coefs.add(coef4)
 
-# Add the coefficient structure coefficient structure
+# Check the coefficient structure for the 4 added times
 #
-# coefs = pyEXP.coefs.SphCoefs(True)
-# coefs.add(coef1)
-# coefs.add(coef2)
-# coefs.add(coef3)
-# coefs.add(coef4)
 print("Times:", coefs.Times())
 
 exit(0)
