### 2024-10-13 8a3d555
Change to pass arrays by copy rather than reference
