### 2024-10-13 34bedce
Update createCoefs.py
Bisect again (?!)
