### 2024-10-13 fa3740c
Update createCoefs.py
add 'sleep()' to try to get some diagnostic output from the runner
