### 2024-10-13 996054d
Update createCoefs.py
