### 2024-10-13 0343410
Update createCoefs.py
