+++ tests/Halo/createCoefs.py
@@ -37,8 +37,6 @@ print("---- created coefficients")
 
 time.sleep(1)
 
-exit(0)
-
 # Call the basis to generate coefficients
 #
 mass = []
