### 2024-10-14 876a923
Rework cuda file handling
Signed-off-by: Georgia Stuart <gstuart@umass.edu>
