+++ expui/CMakeLists.txt
@@ -16,7 +16,7 @@ set(common_INCLUDE $<INSTALL_INTERFACE:include>
 
 
 if(ENABLE_CUDA)
-  list(APPEND common_LINKLIB CUDA::cudart)
+  list(APPEND common_LINKLIB CUDA::toolkit CUDA::cudart)
   if (CUDAToolkit_VERSION VERSION_GREATER_EQUAL 12)
     list(APPEND common_LINKLIB CUDA::nvtx3)
   else ()
