### 2024-10-14 42be626
Add extra checks to CI for runner diagnostic information
