### 2024-10-14 bfbfbc8
Rollback to reference version of setMatrix() and createFromArray(); remove debugging statements from createCoefs.py
