+++ expui/Coefficients.cc
@@ -469,7 +469,7 @@ namespace CoefClasses
     return mat;
   }
   
-  void SphCoefs::setData(double time, Eigen::VectorXcd dat)
+  void SphCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -484,7 +484,7 @@ namespace CoefClasses
     }
   }
   
-  void SphCoefs::setMatrix(double time, Eigen::MatrixXcd dat)
+  void SphCoefs::setMatrix(double time, Eigen::MatrixXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -891,7 +891,7 @@ namespace CoefClasses
     return mat;
   }
   
-  void CylCoefs::setData(double time, Eigen::VectorXcd dat)
+  void CylCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -906,7 +906,7 @@ namespace CoefClasses
     }
   }
 
-  void CylCoefs::setMatrix(double time, Eigen::MatrixXcd dat)
+  void CylCoefs::setMatrix(double time, Eigen::MatrixXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1247,7 +1247,7 @@ namespace CoefClasses
     return arr;
   }
 
-  void SlabCoefs::setData(double time, Eigen::VectorXcd dat)
+  void SlabCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1262,7 +1262,7 @@ namespace CoefClasses
     }
   }
 
-  void SlabCoefs::setTensor(double time, const Eigen3d dat)
+  void SlabCoefs::setTensor(double time, const Eigen3d& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1599,7 +1599,7 @@ namespace CoefClasses
     return arr;
   }
 
-  void CubeCoefs::setData(double time, Eigen::VectorXcd dat)
+  void CubeCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1614,7 +1614,7 @@ namespace CoefClasses
     }
   }
 
-  void CubeCoefs::setTensor(double time, const Eigen3d dat)
+  void CubeCoefs::setTensor(double time, const Eigen3d& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -1968,7 +1968,7 @@ namespace CoefClasses
     return arr;
   }
   
-  void TableData::setData(double time, Eigen::VectorXcd dat)
+  void TableData::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -2458,7 +2458,7 @@ namespace CoefClasses
     return *mat;
   }
   
-  void SphFldCoefs::setData(double time, Eigen::VectorXcd dat)
+  void SphFldCoefs::setData(double time, Eigen::VectorXcd& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -2473,7 +2473,7 @@ namespace CoefClasses
     }
   }
   
-  void SphFldCoefs::setMatrix(double time, SphFldStruct::dataType dat)
+  void SphFldCoefs::setMatrix(double time, SphFldStruct::dataType& dat)
   {
     auto it = coefs.find(roundTime(time));
 
@@ -2897,7 +2897,7 @@ namespace CoefClasses
     return *mat;
   }
   
[…]
