+++ expui/BasisFactory.cc
@@ -272,7 +272,7 @@ namespace BasisClasses
   // Generate coefficients from a phase-space table
   //
   CoefClasses::CoefStrPtr Basis::createFromArray
-  (Eigen::VectorXd m, RowMatrixXd p, double time, std::vector<double> ctr,
+  (Eigen::VectorXd& m, RowMatrixXd& p, double time, std::vector<double> ctr,
    bool roundrobin, bool posvelrows)
   {
     initFromArray(ctr);
