+++ expui/BasisFactory.H
@@ -205,13 +205,13 @@ namespace BasisClasses
 
     //! Accumulate coefficient contributions from arrays
     virtual void
-    addFromArray(Eigen::VectorXd m, RowMatrixXd p, bool roundrobin,
+    addFromArray(Eigen::VectorXd& m, RowMatrixXd& p, bool roundrobin,
 		 bool posvelrows) = 0;
 
     //! Generate coeffients from an array and optional center location
     //! for the expansion
     CoefClasses::CoefStrPtr createFromArray
-    (Eigen::VectorXd m, RowMatrixXd p, double time=0.0,
+    (Eigen::VectorXd& m, RowMatrixXd& p, double time=0.0,
      std::vector<double> center={0.0, 0.0, 0.0},
      bool roundrobin=true, bool posvelrows=false);
 
