### 2024-10-14 92c1f8b
Need at least CMake 3.25 to get the cuda::nvtx3 target
