### 2024-10-14 02d09fb
Relabel test as 'quick' after updating pybind11
