### 2024-10-14 b215b49
Update build.yml
Reverted from `Debug` to `Release`
