### 2024-10-15 ac91538
Update build.yml
test gcc and clang matrix
