+++ .github/workflows/build.yml
@@ -13,13 +13,11 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, ubuntu-22.04]
-        c_compiler: [gcc, clang]
-        cpp_compiler: [g++, clang++]
         include:
-          - c_compiler: gcc
-            cpp_compiler: g++
-          - c_compiler: clang
-            cpp_compiler: clang++
+          - cc: gcc
+            cxx: g++
+          - cc: clang
+            cxx: clang++
 
     name: "Test pyEXP Build"
     runs-on: ${{ matrix.os }}
@@ -44,6 +42,7 @@ jobs:
         if: runner.os == 'Linux'
         env:
           CC: ${{ matrix.cc }}
+          CXX: ${{ matrix.cxx }}
         working-directory: ./build
         run: >-
           cmake
