### 2024-10-15 28fa507
Update build.yml
Need OpenMP support for clang
