### 2024-10-15 60ffa3d
Update CMakeLists.txt
Remove hardwired use of llvm's libc++
