### 2024-10-15 7d9cf63
Decrease of the clang chatter; add 24.04 explicitly to the CI matrix
