### 2024-10-15 136fa16
Update build.yml
Use apt numpy package for 24.04
