### 2024-10-16 0218739
This commit addresses all of the warnings and non-compliance issues identified by Clang and Clang++.  They are all minor.  The result passes 'ctest -L long' but pyEXP-example notebooks have not been run.
