### 2024-10-16 755bd90
Update for 24.04: remove pip install in favor of apt install of numpy
