### 2024-10-16 e712207
custom MacOS fix for fftw_complex destructor problems
