### 2024-10-16 455c03f
Use reinterpret cast from std::complex to handle FFTW complex types
