### 2024-10-22 #7302 @pmocz
Looking good! I had trouble installing it from scratch just because I had missing libraries. The docker script works well. When I try the first tutorial example, I get an error:

```
Cell In[3], line 25
      9 config="""
     10 ---
     11 id: sphereSL
   (...)
     20 ...
     21 """
     23 # Construct the basis instances
     24 #
---> 25 basis = pyEXP.basis.Basis.factory(config)

RuntimeError: SphericalSL requires a specified cachename in your YAML config
for consistency with previous invocations and existing coefficient
sets.  Please add explicitly add 'cachename: name' to your config
with new 'name' for creating a basis or an existing 'name' for
reading a previously generated basis cache
```

which goes away if I specify a `name` in the` config`
