### 2024-10-22 #7302 @pmocz
Don't know if this is helpful, but a visualization or graphic of when inputs and outputs look like somewhere in the intro docs would be helpful to quickly communicate what your code is doing

Saving figures as `.svg` instead of `.png` will also make them look more crisp on the web
