### 2024-10-22 #7302 @pmocz
The Contributing guides says: `cite the EXP published papers`. These should be included in the github repo CITATIONS.bib file
