+++ utils/SL/qtest.cc
@@ -35,7 +35,7 @@ int main(int argc, char** argv)
    ("Lmax", "maximum number of angular harmonics in the expansion",
      cxxopts::value<int>(Lmax)->default_value("2"))
    ("nmax", "maximum number of radial harmonics in the expansion",
-     cxxopts::value<int>(nmax)->default_value("10"))
+     cxxopts::value<int>(nmax)->default_value("18"))
    ("numr", "radial knots for the SL grid",
      cxxopts::value<int>(numr)->default_value("1000"))
    ("rmin", "minimum radius for the SL grid",
