+++ exputil/EmpCyl2d.cc
@@ -836,7 +836,7 @@ void EmpCyl2d::create_tables()
 	for (int l=0; l<nmaxfid; l++) {
 	  D(j, l) += fac * disk->dens(rr) *
 	    basis->potl(m, j, rr) * basis->potl(m, l, rr)
-	    / sqrt(basis->norm(j, m)*basis->norm(l, m)*4.0);
+	    / sqrt(basis->norm(j, m)*basis->norm(l, m));
 	}
       }
     }
@@ -873,15 +873,9 @@ void EmpCyl2d::create_tables()
       if (m==0) xgrid[i] = r;
 
       for (int n=0; n<nmaxfid; n++) {
-<<<<<<< Updated upstream
-	pot(n) = basis->potl(m, n, r) / sqrt(basis->norm(n, m)*0.5);
-	den(n) = basis->dens(m, n, r) / sqrt(basis->norm(n, m)*2.0);
-	dph(n) = basis->dpot(m, n, r) / sqrt(basis->norm(n, m)*0.5);
-=======
-	pot(n) = basis->potl(m, n, r) / sqrt(basis->norm(n, m)*0.5/M_PI);
-	den(n) = basis->dens(m, n, r) / sqrt(basis->norm(n, m)*2.0*M_PI);
-	dph(n) = basis->dpot(m, n, r) / sqrt(basis->norm(n, m)*0.5/M_PI);
->>>>>>> Stashed changes
+	pot(n) = basis->potl(m, n, r) / sqrt(basis->norm(n, m));
+	den(n) = basis->dens(m, n, r) / sqrt(basis->norm(n, m));
+	dph(n) = basis->dpot(m, n, r) / sqrt(basis->norm(n, m));
       }
       
       pot = U.transpose() * pot;
@@ -1209,11 +1203,7 @@ double EmpCyl2d::get_potl(double r, int M, int N)
   checkMN(M, N, "get_potl");
 
   if (basis_test) {
-<<<<<<< Updated upstream
-    return basis->potl(M, N, r)/sqrt(basis->norm(N, M)*0.5);
-=======
-    return basis->potl(M, N, r)/sqrt(basis->norm(N, M)*0.5/M_PI);
->>>>>>> Stashed changes
+    return basis->potl(M, N, r)/sqrt(basis->norm(N, M));
   }
 
   int lo, hi;
@@ -1228,11 +1218,7 @@ double EmpCyl2d::get_dens(double r, int M, int N)
   checkMN(M, N, "get_dens");
 
   if (basis_test) {
-<<<<<<< Updated upstream
-    return basis->dens(M, N, r)/sqrt(basis->norm(N, M)*2.0);
-=======
-    return basis->dens(M, N, r)/sqrt(basis->norm(N, M)*2.0*M_PI);
->>>>>>> Stashed changes
+    return basis->dens(M, N, r)/sqrt(basis->norm(N, M));
   }
 
   int lo, hi;
@@ -1247,11 +1233,7 @@ double EmpCyl2d::get_dpot(double r, int M, int N)
   checkMN(M, N, "get_dpot");
 
   if (basis_test) {
-<<<<<<< Updated upstream
-    return basis->dpot(M, N, r)/sqrt(basis->norm(N, M)*0.5);
-=======
-    return basis->dpot(M, N, r)/sqrt(basis->norm(N, M)*2.0*M_PI);
->>>>>>> Stashed changes
+    return basis->dpot(M, N, r)/sqrt(basis->norm(N, M));
   }
 
   int lo, hi;
@@ -1325,11 +1307,7 @@ void EmpCyl2d::checkCoefs()
     
       for (int j=0; j<nmax; j++) {
[…]
