+++ pyEXP/TensorToArray.H
@@ -27,37 +27,6 @@ py::array_t<T> make_ndarray3(Eigen::Tensor<T, 3>& mat)
      );
 }
 
-template <typename T>
-Eigen::Tensor<T, 3> make_tensor3(py::array_t<T> array)
-{
-  // Request a buffer descriptor from Python
-  py::buffer_info buffer_info = array.request();
-
-  // Get the array dimenions
-  T *data = static_cast<T *>(buffer_info.ptr);
-  std::vector<ssize_t> shape = buffer_info.shape;
-
-  // Check rank
-  if (shape.size() != 3) {
-    std::ostringstream sout;
-    sout << "make_tensor3: tensor rank must be 3, found "
-	 << shape.size();
-    throw std::runtime_error(sout.str());
-  }
-
-  // Build result tensor with col-major ordering
-  Eigen::Tensor<T, 3> tensor(shape[0], shape[1], shape[2]);
-  for (int i=0, l=0; i < shape[0]; i++) {
-    for (int j=0; j < shape[1]; j++) {
-      for (int k=0; k < shape[2]; k++) {
-	tensor(i, j, k) = data[l++];
-      }
-    }
-  }
-
-  return tensor;
-}
-
 
 //! Helper function that maps the Eigen::Tensor<T, 4> into an numpy.ndarray
 template <typename T>
