### 2024-10-28 2752d2e
Fix clang options for Linux
