+++ CMakeLists.txt
@@ -16,8 +16,6 @@ if(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   message(STATUS "Clang compiler version is: ${CMAKE_CXX_COMPILER_VERSION}")
   if(CMAKE_SYSTEM_NAME MATCHES "Linux")
     if(CMAKE_CXX_COMPILER_VERSION VERSION_LESS 18.0.0)
-      set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -stdlib=libc++")
-    else()
       set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -stdlib=libstdc++")
     endif()
   endif()
