### 2024-11-02 7443379
Make the force/coefficient evaluation consistent with the 2d basis norm
