### 2024-11-06 321dcd8
Updates for standalone VTK file writing
