### 2024-11-12 203cade
Add KDtree to its own namespace and allow a NN cache for multithreaded operation
