+++ include/KDtree.H
@@ -6,381 +6,414 @@
 #include <tuple>
 #include <map>
 
-//! Wrapper class for std::map to keep a finite size
-template<typename key_t, typename value_t>
-class Cache : public std::map<key_t, value_t>
+namespace KDtree
 {
-public:
-  Cache(int N=1) : max_size(N) {}
-
-  void resize(int N) { max_size = N; }
-
-  void add( key_t key, value_t value )
+  //! Wrapper class for std::map to keep a finite size
+  template<typename key_t, typename value_t>
+  class Cache : public std::map<key_t, value_t>
   {
-    (*this)[key] = value;
-    if (this->size() > max_size) {
-      this->erase(--(*this).end());
-    }
-  }
+  public:
+    Cache(int N=1) : max_size(N) {}
 
-private:
-  int max_size;
-};
+    void resize(int N) { max_size = N; }
 
-
-/** Class for representing a point
-    Coordinate_type must be a numeric type
-    Field (weight) is a double
-*/
-template<typename coordinate_type, size_t dimensions>
-class point
-{
-public:
-  //! Null initializer
-  point() {}
-  
-  //! Constructor
-  point(std::array<coordinate_type, dimensions> c, double m=1,
-	unsigned long indx=0) : coords_(c), mass_(m), indx_(indx) {}
-  
-  point(std::array<coordinate_type, dimensions> c,
-	std::array<coordinate_type, dimensions> v,
-	double m=1, unsigned long indx=0) : coords_(c), vels_(v),
-					    mass_(m), indx_(indx) {}
-  
-  //! List constructor
-  point(std::initializer_list<coordinate_type> list, double m=1,
-	unsigned long indx=0) : mass_(m), indx_(indx)
-  {
-    size_t n = std::min(dimensions, list.size());
-    std::copy_n(list.begin(), n, coords_.begin());
-  }
-
-  point(std::initializer_list<coordinate_type> list,
-	std::initializer_list<coordinate_type> vlst,
-	double m=1,
-	unsigned long indx=0) : mass_(m), indx_(indx)
-  {
-    size_t n = std::min(dimensions, list.size());
-    std::copy_n(list.begin(), n, coords_.begin());
-    std::copy_n(vlst.begin(), n, vels_.begin());
-  }
-
-  //! Copy constructor
-  point(const point& p)
-  {
-    for (size_t n=0; n<dimensions; n++) coords_[n] = p.coords_[n];
-    for (size_t n=0; n<dimensions; n++) vels_  [n] = p.vels_  [n];
-    mass_ = p.mass_;
-    indx_ = p.indx_;
-  }
-
-  /**
-   * Returns the coordinate in the given dimension.
-   *
-   * @param index dimension index (zero based)
-   * @return coordinate in the given dimension
-   */
-  coordinate_type get(size_t index) const
-  {
-    return coords_[index];
-  }
-  
-  coordinate_type vel(size_t index) const
-  {
-    return vels_[index];
-  }
-  
-  /**
-   * Returns the distance squared from this point to another
-   * point.
-   *
-   * @param pt another point
-   * @return distance squared from this point to the other point
-   */
-  double distance(const point& pt) const
-  {
-    double dist = 0;
-    for (size_t i = 0; i < dimensions; ++i)
-      {
-	double d = get(i) - pt.get(i);
-	dist += d * d;
-      }
-    return dist;
-  }
-
-  /**
-   * Returns the squared velocity between this point to another
-   * point.
-   *
-   * @param pt another point
-   * @return squared speed between this point and the other point
-   */
-  double speed(const point& pt) const
-  {
-    double rvel = 0;
[…]
