### 2024-11-15 6817c59
Update to normalization scheme for 2-d bases
