+++ src/PolarBasis.cc
@@ -401,8 +401,8 @@ void * PolarBasis::determine_coefficients_thread(void * arg)
 {
   // For biorthogonal density component and normalization
   // 
-  constexpr double norm0 = 0.5*M_2_SQRTPI/M_SQRT2;
-  constexpr double norm1 = 0.5*M_2_SQRTPI;
+  constexpr double norm0 = 1.0;
+  constexpr double norm1 = M_SQRT2;
 
   double r, r2, facL=1.0, fac1, fac2, phi, mass;
   double xx, yy, zz;
@@ -1044,8 +1044,8 @@ void PolarBasis::multistep_update(int from, int to, Component *c, int i, int id)
 
   // For biorthogonal density component and normalization
   // 
-  constexpr double norm0 = 0.5*M_2_SQRTPI/M_SQRT2;
-  constexpr double norm1 = 0.5*M_2_SQRTPI;
+  constexpr double norm0 = 1.0;
+  constexpr double norm1 = M_SQRT2;
 
   double mass = c->Mass(i) * component->Adiabatic();
 
@@ -1323,8 +1323,8 @@ void PolarBasis::compute_multistep_coefficients()
 
 void * PolarBasis::determine_acceleration_and_potential_thread(void * arg)
 {
-  constexpr double norm0 = 0.5*M_2_SQRTPI/M_SQRT2;
-  constexpr double norm1 = 0.5*M_2_SQRTPI;
+  constexpr double norm0 = 1.0;
+  constexpr double norm1 = M_SQRT2;
 
   double r, r0=0.0, phi;
   double potr, potz, potl, potp, p, pc, drc, drs, dzc, dzs, ps, dfacp, facdp;
