### 2024-11-15 c245d79
Fix backward compatibility for old GNU and except Clang
