### 2024-11-19 f9a0f20
Separate background from target models in PolarModel
