+++ src/FlatDisk.cc
@@ -21,25 +21,33 @@ FlatDisk::valid_keys = {
   "model",
   "biorth",
   "diskconf",
-  "cachename"
+  "background",
+  "cachename",
+  "dump"
 };
 
 FlatDisk::FlatDisk(Component* c0, const YAML::Node& conf, MixtureBasis* m) :
   PolarBasis(c0, conf, m)
 {
   id = "FlatDisk EOF";
-				// Defaults
+
+  // Defaults
+  //
   knots      = 40;
   numr       = 2000;
   rcylmin    = 0.0;
   rcylmax    = 10.0;
   logr       = false;
   is_flat    = true;
+  dump_basis = false;
 
-				// Get initialization info
+  // Get initialization info
+  //
   initialize();
 
-  id += "[" + model + "][" + biorth + "]";
+  // Append info to basis ID
+  //
+  id += "[" + ortho->getModelName() + "][" + ortho->getBiorthName() + "]";
 
   if (myid==0) {
     std::string sep("----    ");
@@ -81,6 +89,8 @@ void FlatDisk::initialize()
     if (conf["biorth"])    biorth     = conf["biorth"].as<std::string>();
 
     if (conf["mmax"]) Lmax = Mmax = mmax; // Override base-class values
+
+    if (conf["dump"])     dump_basis  = conf["dump"].as<bool>();
   }
   catch (YAML::Exception & error) {
     if (myid==0) std::cout << "Error parsing parameters in FlatDisk: "
@@ -96,11 +106,92 @@ void FlatDisk::initialize()
   // Create the BiorthCyl instance
   ortho = std::make_shared<BiorthCyl>(conf);
 
+  if (dump_basis) ortho->dump_basis(runtag);
+
+  // Set background model
+  if (conf["background"]) setBackground();
 }
 
 FlatDisk::~FlatDisk(void)
 {
   // NADA
+}
+
+
+void FlatDisk::setBackground()
+{
+  try {
+
+    YAML::Node Params = conf["background"];
+
+    std::string name = Params["name"].as<std::string>();
+    auto params      = Params["parameters"];
+    
+    // Convert ID string to lower case
+    //
+    std::transform(name.begin(), name.end(), name.begin(),
+		   [](unsigned char c){ return std::tolower(c); });
+    
+    if (name.find("kuzmin") != std::string::npos) {
+      if (myid==0) {
+	std::cout << "---- FlatDisk: Making a Kuzmin disk";
+	if (params) std::cout << " with" << std::endl << Params["parameters"];
+	std::cout << std::endl;
+      }
+      disk = std::make_shared<EmpCyl2d::KuzminCyl>(params);
+      return;
+    }
+      
+    if (name.find("mestel") != std::string::npos) {
+      if (myid==0) {
+	std::cout << "---- FlatDisk: Making a finite Mestel disk";
+	if (params) std::cout << " with" << std::endl << Params["parameters"];
+	std::cout << std::endl;
+      }
+      disk = std::make_shared<EmpCyl2d::MestelCyl>(params);
+      return;
+    }
+      
+    if (name.find("zang") != std::string::npos) {
+      if (myid==0) {
+	std::cout << "---- FlatDisk: Making a double-tapered Zang";
+	YAML::Emitter out;
+	out << YAML::Flow << params;
+	if (params) std::cout << " with " << out.c_str();
+	std::cout << std::endl;
+      }
+      disk = std::make_shared<EmpCyl2d::ZangCyl>(params);
+      return;
+    }
+      
+    if (name.find("expon") != std::string::npos) {
+      if (myid==0) {
+	std::cout << "---- FlatDisk: Making an Exponential disk";
+	if (params) std::cout << " with" << std::endl << params;
[…]
