### 2024-11-19 cd0a276
Fixes for FlatDisk with CUDA
