+++ src/cudaBiorthCyl.cu
@@ -193,9 +193,8 @@ void BiorthCyl::initialize_cuda
 
   for (int i=0; i<numr; i++) {
     double r = xi_to_r(xmin + dx0*i);
-    auto [p, dr, d] = disk->background(r);
-    tt[0][i] = p;
-    tt[1][i] = dr;
+    tt[0][i] =  disk->pot(r);
+    tt[1][i] = -disk->dpot(r);
   }
 
   // Allocate CUDA array in device memory (a one-dimension 'channel')
