### 2024-11-20 817892c
Added CBDisk for testing; not optimized in any way and no CUDA implementation
