+++ src/CBDisk.cc
@@ -1,0 +1,510 @@
+#include <limits>
+
+#include <Eigen/Eigen>
+
+#include "expand.H"
+
+#include <gaussQ.H>
+#include <CBDisk.H>
+#include <interp.H>
+
+const std::set<std::string>
+CBDisk::valid_keys = {
+  "mmax",
+  "scale",
+  "diskconf",
+  "background"
+};
+
+CBDisk::CBDisk(Component* c0, const YAML::Node& conf, MixtureBasis* m) :
+  PolarBasis(c0, conf, m)
+{
+  id = "CBDisk";
+
+  // Radial scale factor/softening radius
+  //
+  scale = 1.0;
+
+  // Get initialization info
+  //
+  initialize();
+
+  if (myid==0) {
+    std::string sep("----    ");
+    std::cout << "---- CBDisk parameters: "
+	      << std::endl << sep << "scale="       << scale
+	      << std::endl << sep << "lmax="        << Lmax
+	      << std::endl << sep << "nmax="        << nmax
+	      << std::endl << sep << "NO_L0="       << std::boolalpha << NO_M0
+	      << std::endl << sep << "NO_L1="       << std::boolalpha << NO_M1
+	      << std::endl << sep << "EVEN_M="      << std::boolalpha << EVEN_M
+	      << std::endl << sep << "M0_ONLY="     << std::boolalpha << M0_only
+	      << std::endl << sep << "selfgrav="    << std::boolalpha << self_consistent
+	      << std::endl;
+  }
+
+  setup();
+  if (myid==0) orthoCheck();
+}
+
+
+void CBDisk::initialize()
+{
+  // Remove matched keys
+  //
+  for (auto v : valid_keys) current_keys.erase(v);
+  
+  // Assign values from YAML
+  //
+  try {
+    if (conf["mmax"])      mmax       = conf["mmax"].as<int>();
+    if (conf["Lmax"])      mmax       = conf["Lmax"].as<int>();
+    if (conf["Mmax"])      mmax       = conf["Mmax"].as<int>();
+    if (conf["scale"])     scale      = conf["scale"].as<double>();
+
+    Lmax = Mmax = mmax;		// Override base-class values
+  }
+  catch (YAML::Exception & error) {
+    if (myid==0) std::cout << "Error parsing parameters in CBDisk: "
+			   << error.what() << std::endl
+			   << std::string(60, '-') << std::endl
+			   << "Config node"        << std::endl
+			   << std::string(60, '-') << std::endl
+			   << conf                 << std::endl
+			   << std::string(60, '-') << std::endl;
+    throw std::runtime_error("CBDisk::initialize: error in parsing YAML");
+  }
+
+  // Set background model
+  if (conf["background"]) setBackground();
+}
+
+CBDisk::~CBDisk(void)
+{
+  // NADA
+}
+
+
+void CBDisk::setBackground()
+{
+  try {
+
+    YAML::Node Params = conf["background"];
+
+    std::string name = Params["name"].as<std::string>();
+    auto params      = Params["parameters"];
+    
+    // Convert ID string to lower case
+    //
+    std::transform(name.begin(), name.end(), name.begin(),
+		   [](unsigned char c){ return std::tolower(c); });
+    
+    if (name.find("kuzmin") != std::string::npos) {
[…]
