### 2024-11-21 18aa52b
Zero background arrays and only fill if disk object is instantiated
