### 2024-11-21 6e5dbe4
Allow 2d and 3d norm switch; only 2d currently used so this makes no change
