### 2024-11-21 fac5609
Output updates only
