+++ expui/BiorthBasis.cc
@@ -1523,6 +1523,7 @@ namespace BasisClasses
     "M0_ONLY",
     "NO_MONO",
     "diskconf",
+    "background",
     "ssfrac",
     "playback",
     "coefMaster",
@@ -1953,7 +1954,7 @@ namespace BasisClasses
     return {v[0], v[1], v[2], v[3], v[4], v[5], potx, poty, v[7]};
   }
 
-  std::vector<Eigen::MatrixXd> FlatDisk::orthoCheck()
+  std::vector<Eigen::MatrixXd> FlatDisk::orthoCheck(int knots)
   {
     return ortho->orthoCheck();
   }
@@ -1984,6 +1985,657 @@ namespace BasisClasses
       ortho->get_pot   (tabpot, pow(10.0, logxmin + dx*i), 0.0);
       ortho->get_dens  (tabden, pow(10.0, logxmin + dx*i), 0.0);
       ortho->get_rforce(tabrfc, pow(10.0, logxmin + dx*i), 0.0);
+      for (int m=0; m<=mmax; m++) {
+	for (int n=0; n<nmax; n++){
+	  ret[m][n]["potential"](i) = tabpot(m, n);
+	  ret[m][n]["density"  ](i) = tabden(m, n);
+	  ret[m][n]["rforce"   ](i) = tabrfc(m, n);
+	}
+      }
+    }
+    
+    return ret;
+  }
+
+  const std::set<std::string>
+  CBDisk::valid_keys = {
+    "self_consistent",
+    "NO_M0",
+    "NO_M1",
+    "EVEN_M",
+    "M0_BACK",
+    "M0_ONLY",
+    "NO_MONO",
+    "background",
+    "playback",
+    "coefMaster",
+    "scale",
+    "Lmax",
+    "Mmax",
+    "nmax",
+    "mmax",
+    "mlim",
+    "dof",
+    "subsamp",
+    "samplesz",
+    "vtkfreq",
+    "tksmooth",
+    "tkcum",
+    "tk_type"
+  };
+
+  CBDisk::CBDisk(const YAML::Node& CONF) :
+    BiorthBasis(CONF, "CBDisk")
+  {
+    initialize();
+  }
+
+  CBDisk::CBDisk(const std::string& confstr) :
+    BiorthBasis(confstr, "CBDisk")
+  {
+    initialize();
+  }
+
+  void CBDisk::initialize()
+  {
+
+    // Assign some defaults
+    //
+    mmax       = 6;
+    nmax       = 18;
+    scale      = 1.0;
+    
+    // Check for unmatched keys
+    //
+    auto unmatched = YamlCheck(conf, valid_keys);
+    if (unmatched.size())
+      throw YamlConfigError("Basis::Basis::CBDisk", "parameter", unmatched, __FILE__, __LINE__);
+    
+    // Assign values from YAML
+    //
+    try {
+      if (conf["Lmax"])      mmax       = conf["Lmax"].as<int>(); // Proxy
+      if (conf["Mmax"])      mmax       = conf["Mmax"].as<int>(); // Proxy
+      if (conf["mmax"])      mmax       = conf["mmax"].as<int>();
+      if (conf["nmax"])      nmax       = conf["nmax"].as<int>();
+      if (conf["scale"])     scale      = conf["scale"].as<double>();
+      
+      N1 = 0;
+      N2 = std::numeric_limits<int>::max();
+      NO_M0 = NO_M1 = EVEN_M = M0_only = false;
+      
+      if (conf["N1"]   )     N1        = conf["N1"].as<bool>();
[…]
