### 2024-11-26 9ec4f4a
Allow gendisk2d to create an arbitrary EmpCyl2d basis
