### 2024-11-26 5a51669
Fix scaling for CB disk
