+++ src/CBDisk.cc
@@ -45,6 +45,12 @@ CBDisk::CBDisk(Component* c0, const YAML::Node& conf, MixtureBasis* m) :
   }
 
   setup();
+
+  // Potential, force, and density scaling
+  //
+  fac1 = pow(scale, -0.5);
+  fac2 = pow(scale, -1.5);
+
   if (myid==0) orthoCheck();
 }
 
@@ -178,8 +184,8 @@ void CBDisk::get_dpotl(double r, double z,
 
   for (int m=0; m<=mmax; m++) {	// Pack the array
     for (int j=0; j<nmax; j++) {
-      pot(m, j) = potl(j, m, R) / scale;
-      dpr(m, j) = dpot(j, m, R) / (scale*scale);
+      pot(m, j) = potl(j, m, R) * fac1;
+      dpr(m, j) = -dpot(j, m, R) * fac1 / scale;
     }
   }
 }
@@ -192,7 +198,7 @@ void CBDisk::get_potl(double r, double z, Eigen::MatrixXd& p, int tid)
 
   for (int m=0; m<=mmax; m++) {	// Pack the array
     for (int j=0; j<nmax; j++)
-      p(m, j) = potl(j, m, R)/ scale;
+      p(m, j) = potl(j, m, R) * fac1;
   }
 }
 
@@ -204,7 +210,7 @@ void CBDisk::get_dens(double r, double z, Eigen::MatrixXd& p, int tid)
 
   for (int m=0; m<=mmax; m++) {	// Pack the array
     for (int j=0; j<nmax; j++)
-      p(m, j) = dens(j, m, R) * scale*scale;
+      p(m, j) = dens(j, m, R) * fac2;
   }
 }
 
@@ -215,12 +221,12 @@ void CBDisk::get_potl_dens(double r, double z,
   p.resize(mmax+1, nmax);	// Resize the return arrays
   d.resize(mmax+1, nmax);
 
-  double R = r/scale;
+  double R = r/scale, facp = pow(scale, -0.5), facd = pow(scale, -1.5);
 
   for (int m=0; m<=mmax; m++) {	// Pack the array
     for (int j=0; j<nmax; j++) {
-      p(m, j) = potl(j, m, R) / scale;
-      d(m, j) = dens(j, m, R) * scale*scale;
+      p(m, j) = potl(j, m, R) * facp;
+      d(m, j) = dens(j, m, R) * facd;
     }
   }
 }
@@ -391,32 +397,40 @@ void CBDisk::orthoCheck()
   const int num = 4000;
   LegeQuad lq(num);
 
-  Eigen::MatrixXd ortho(nmax, nmax);
+  std::vector<Eigen::MatrixXd> ortho(mmax+1);
   Eigen::VectorXd worst(mmax+1);
+  Eigen::MatrixXd vpot(mmax+1, nmax), vden(mmax+1, nmax);
 
   double Rmax = scale*100.0;
 
-
-  // Direct test
-  //
   worst.setZero();
 
-  for (int m=0; m<=mmax; m++) {
+  // Allocate result matrices
+  //
+  for (auto & v : ortho) {
+    v.resize(nmax, nmax);
+    v.setZero();
+  }
 
-    ortho.setZero();
+  for (int i=0; i<num; i++) {
+    double r = lq.knot(i) * Rmax, fac = lq.weight(i) * Rmax;
 
-    for (int i=0; i<num; i++) {
-      double r = lq.knot(i) * Rmax, fac = lq.weight(i) * Rmax;
+    get_potl(r, 0.0, vpot, 0);
+    get_dens(r, 0.0, vden, 0);
+
+    for (int m=0; m<=mmax; m++) {
[…]
