### 2024-11-26 57d477b
Add scaling for CBDisk to pyEXP
