### 2024-12-08 6df822f
Remove the default 'diag' output suffix
