### 2024-12-09 EXP-code/EXP#94 @schuhmaj
[JOSS REVIEW] Improve Install Instructions/ Process
# The issue

JOSS requires

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

This bullet point is **fulfilled**, so this is not a blocking review criteria. However, I do have some suggestions to make `EXP` easier to set up.

### Suggestions to improve installation instructions

- The installation's dependencies are not fully stated in the [INSTALL.md](https://github.com/EXP-code/EXP/blob/main/INSTALL.md). I mainly refer to the requirement to have the `fftw` or `eigen3` library pre-installed
- The [installation guide](https://exp-docs.readthedocs.io/en/latest/intro/install.html) should ideally start with the requirements. As a new user, I am basically executing it step-by-step. However, the prerequisites (which are not automatically set-up via git submodules) `eigen3`, `fftw` are listed *after* the actual configuration with CMake. I would recommend to re-order this.

### Suggestions to improve CMake Build System overall

Alternatively, I would recommend embracing the existing CMake build system more:
- You can set up the [FFTW](http://fftw.org/download.html) or [HDF5](https://www.hdfgroup.org/download-hdf5/source-code/) (and the other) libraries via CMake's [Fetch Content Module](https://cliutils.gitlab.io/modern-cmake/chapters/projects/fetch.html). Both libraries offer tar balls containing a CMakeList.txt. E.g.

```cmake
FetchContent_Declare(
  fftw
  URL      http://fftw.org/fftw-3.3.10.tar.gz
  URL_HASH MD5=8ccbf6a5ea78a16dbc3e1306e234cc5c
)
```

----
This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7302
