### 2024-12-09 EXP-code/EXP#95 @schuhmaj
[JOSS REVIEW] Outdated Links/ Descriptions
# The issue

Also, a minor thing (i.e., this is not a blocking criterion in the context of the JOSS Review.), the documentation contains some outdated/ broken links. E.g.:

- The link to the HDF5 library in the [install guide](https://exp-docs.readthedocs.io/en/latest/intro/install.html): Faulty Link https://portal.hdfgroup.org/display/HDF5/HDF5
- The [EXP Tutorial Page](https://exp-docs.readthedocs.io/en/latest/intro/EXP-tutorial.html) doesn't correctly link to the example repository: Faulty Link https://github.com/orgs/EXP-examples

The example repositories READMEs are sometimes out-of-date. E.g.
- [Here](https://github.com/EXP-code/pyEXP-examples/tree/main/How-To/Recipes), the `README.md` lists bullet points which don't exist (anymore?)

----
This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7302
