### 2024-12-12 EXP-code/EXP#97 @schuhmaj
[JOSS REVIEW] Add comparison to commonly used N-Body Codes into the Paper
# The issue

JOSS requires

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

While your current JOSS paper includes references to the employed technologies. A comparison/ mentioning of software offering similar functionality as `EXP` is missing.

I've read a bit through your references. In Petersen et al. [1] you mention `GADGET-2`. Hence, this issue should be fairly trivial to resolve by adding a few sentences how `EXP` compares to `GADGET-2` and other N-body codes in a paragraph of the JOSS paper.

[1] Michael S Petersen, Martin D Weinberg, Neal Katz, EXP: N-body integration using basis function expansions, Monthly Notices of the Royal Astronomical Society, Volume 510, Issue 4, March 2022, Pages 6201–6217, https://doi.org/10.1093/mnras/stab3639

----
This issue is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/7302
