### 2024-12-13 #7302 @schuhmaj
Hi - I concluded my review and I fully recommend accepting this submission into JOSS. The software is amazing. There are only two small improvements which need to be undertaken upfront so that I can checkmark the last two items from the list.

- [ ] https://github.com/EXP-code/EXP/issues/97
  - The paper does currently not compare its implementation to other well-known packages in the area, but restricts itself to presenting only dependencies.
  - Given that such a comparison exists in a referenced paper by the same authors, this should be low-effort to add:
- [ ] https://github.com/EXP-code/EXP/issues/96
  - The dependencies' versions are not documented to a satisfactory level. I had a compile error with`eigen 3.3.7` on one of my three testing systems, only resolvable switching to `eigen 3.4`
  - Resolvable by either specifying the concrete versions or by embracing the CMake build system more to make the installation fully automatic (relates to https://github.com/EXP-code/EXP/issues/94)

The two other opened issues are strong recommendations, but no blockers in the context of this review.
