### 2024-12-17 febd3b7
Added a spherical histogram1d with logarithmic scaling to pyEXP; since this is a new function, there is no API change
