### 2024-12-18 1205a0a
Code clean up
This partially reverts commit 2336b9af5777c754e12596b0f43827c1ca2a3d31.

Updated sanity check to allow negative tstep
