### 2024-12-18 EXP-code/EXP#96 @michael-petersen
The eigen<3.4 issue just bit me in a different fresh compile as well -- one strategy around this is to submodule Eigen (and potentially also hdf5), so that we can control the version explicitly. The downside is that some extra code is added, and we'll (possibly) have to manually manage the versions.
