### 2024-12-19 46e7b2e
Implement pseudo-forces for 'Orient' moving 'CENTER' non-inertial frames
