+++ src/Orient.H
@@ -8,6 +8,7 @@
 
 #include <euler.H>
 #include <Particle.H>
+#include <PseudoAccel.H>
 
 #include <config_exp.h>
 
@@ -96,7 +97,7 @@ private:
   int keep, current;
   double damp;
   Eigen::Matrix3d body, orig;
-  Eigen::Vector3d axis, center, axis1, center1, center0, cenvel0;
+  Eigen::Vector3d axis, center, axis1, center1, center0, cenvel0, pseudo;
   double sumX, sumX2;
   Eigen::Vector3d sumY, sumY2, sumXY, slope, intercept;
   double sigA, sigC, sigCz;
@@ -110,8 +111,10 @@ private:
   string logfile;
   bool linear;
   
-  vector<double> pos, psa, vel;
+  std::vector<double> pos, psa, vel;
   
+  std::shared_ptr<PseudoAccel> accel;
+
   void accumulate_cpu(double time, Component* c);
 #if HAVE_LIBCUDA==1
   void accumulate_gpu(double time, Component* c);
@@ -125,7 +128,7 @@ public:
   enum ControlFlags {DIAG=1, KE=2, EXTERNAL=4};
   
   //! Constructor
-  Orient(int number_to_keep, int target,
+  Orient(int number_to_keep, int target, int Naccel,
 	 unsigned orient_flags, unsigned control_flags, string logfile,
 	 double dt=0.0, double damping=1.0);
   
@@ -167,6 +170,14 @@ public:
   //! Return current center
   Eigen::Vector3d& currentCenter(void) {return center;};
   
+  //! Return current pseudo force
+  Eigen::Vector3d& currentAccel(void)
+  {
+    pseudo.setZero();
+    if (accel) return pseudo = accel->accel();
+    return pseudo;
+  }
+
   //! Return variance of axis determination (linear least squares solution)
   double currentAxisVar(void) {return sigA;}
   
