### 2024-12-20 46a36df
Initialize Component's 'modified' value
