### 2024-12-20 d2f5f5c
Make debug value static to prevent namespace and addressing issues
