### 2024-12-21 64c87cc
Removed RKHS stuff and made a new branch for the rest
