+++ expui/LiouvilleRKHS.H
@@ -1,284 +1,0 @@
-#ifndef EXP_LIOUVILLE_RKHS_H
-#define EXP_LIOUVILLE_RKHS_H
-
-#include <yaml-cpp/yaml.h>
-#include "CoefContainer.H"
-
-#include <map>
-
-namespace MSSA
-{
-  /**
-     Class for eDMD processing of trajectories using Liouville opertor
-     RKHS.  This implements both Algorithms 6.1 and 7.1 from the paper
-     by Joel A. Rosenfeld and Rushikesh Kamalapurkar, "Singular
-     Dynamic Mode Decomposition", 2023, SIAM J. Appl. Dynamical
-     Systems, Vol. 22, No. 3, pp 2357-2381.
-  */
-  class LiouvilleRKHS
-  {
-
-  public:
-
-    //! RKHS types
-    enum class RKHS
-    {
-      Polynomial,
-      Exponential,
-      Gaussian
-    };
-   
-    //! Analysis method
-    enum class Method
-    {
-      Singular,
-      Eigenfunction
-    };
-   
-    //! For RKHS type reflection and parsing
-    static std::map<RKHS, std::string> RKHS_names;
-    static std::map<std::string, RKHS> RKHS_values;
-
-    //! For method type reflection and parsing
-    static std::map<Method, std::string> Method_names;
-    static std::map<std::string, Method> Method_values;
-
-  protected:
-
-    //@{
-    //! Repacked stream data for eDMD
-    std::map<Key, Eigen::MatrixXd, mSSAkeyCompare > data;
-    //@}
-
-    //! Coefficient container
-    CoefContainer coefDB;
-
-    //! Working updated copy of DB
-    std::shared_ptr<CoefContainer> newDB;
-
-    //! Parameter database
-    YAML::Node params;
-
-    //! Method
-    Method method = Method::Eigenfunction;
-
-    //! Primary Liouville analysis
-    void analysis()
-    {
-      switch (method)
-	{
-	case Method::Singular:
-	  singular_analysis();
-	  break;
-	case Method::Eigenfunction:
-	  eigenfunction_analysis();
-	  break;
-	}
-    }
-
-    //! Singular Louville analysis
-    void singular_analysis();
-
-    //! Eigenfunction Liouville analysis
-    void eigenfunction_analysis();
-
-    //! Compute G matrix for RKHS space defined by mu value
-    Eigen::MatrixXd computeGrammian(double mu, double rat=1.0);
-
-    //! Compute occupation-kernel difference matrix
-    Eigen::MatrixXd computeGammaDiff(double mu);
-
-    //! Compute occupation-kernel vector
-    Eigen::VectorXd computeGamma(const Eigen::VectorXd& x, double mu);
-
-    //! Compute singular Liouville RKHS matrix
-    Eigen::MatrixXd computeGramGamma(double mu);
-
-    //! Compute singular Liouville trajectory matrix
-    Eigen::MatrixXd trajectory();
-
-    bool computed, reconstructed;
-
-    //@{
-    //! Parameters
-    double mu1=1.0, mu2=1.1, eps=1.0e-4;
-    bool use_red = true;
-    //@}
-    
-    //! DMD state matrices
-    Eigen::MatrixXd G1, G2, G3;
-
-    //! Occupation RKHS matrix
-    Eigen::MatrixXd A;
-
-    //! Occupation kernel matrix
-    Eigen::MatrixXd O;
-
-    //! Liouville modes for eigenfunction method
-    Eigen::MatrixXcd Xi;
-
-    //! Liouville modes for singular method
-    Eigen::MatrixXd Xh;
-
-    //! Eigenbasis
[…]
