+++ expui/KoopmanRKHS.cc
@@ -1,840 +1,0 @@
-#define TESTING
-//
-// EDMD (Koopman theory) for EXP coefficients
-//
-// Uses fixed-rank approximation for the SVD to save time and space.
-// Uses the approximate SVD randomized algorithm from Halko,
-// Martinsson, and Tropp by default.  Use BDCSVD or Jacobi flags for
-// the standard methods.  Using the randomized method together with
-// trajectory matrix rather than covariance matrix analysis may lead
-// to large errors in eigenvectors.
-//
-// The implementation here is based on the
-//
-// M. O. Williams, C. W. Rowley, I. G. Kevrekidis, 2015, "A
-// kernel-based method for data-driven Koopman spectral analysis",
-// Journal of Computational Dynamics, 2 (2), 247-265
-//
-
-#include <filesystem>
-#include <stdexcept>
-#include <iostream>
-#include <iomanip>
-#include <fstream>
-#include <sstream>
-#include <vector>
-#include <limits>
-#include <cmath>
-#include <map>
-
-#include <config_exp.h>
-
-#include <Eigen/Dense>
-
-#include <highfive/highfive.hpp>
-#include <highfive/eigen.hpp>
-
-/* For debugging
-   #undef eigen_assert
-   #define eigen_assert(x)						\
-   if (!(x)) { throw (std::runtime_error("Eigen assert error")); }
-*/
-
-#include <omp.h>
-
-#include <KoopmanRKHS.H>
-
-#include <RedSVD.H>
-#include <YamlConfig.H>
-#include <YamlCheck.H>
-#include <EXPException.H>
-#include <libvars.H>
-
-namespace MSSA {
-
-  // Helper ostream manipulator for debug coefficient key info
-  //
-  std::ostream& operator<< (std::ostream& out, const std::vector<unsigned>& t);
-
-  // Get RKHS name from enum
-  //
-  std::map<KoopmanRKHS::RKHS, std::string> KoopmanRKHS::RKHS_names =
-    {
-      {KoopmanRKHS::RKHS::Polynomial,  "Polynomial" },
-      {KoopmanRKHS::RKHS::Exponential, "Exponential"},
-      {KoopmanRKHS::RKHS::Gaussian,    "Gaussian"   }
-    };
-
-  // Get RKHS enum from name
-  //
-  std::map<std::string, KoopmanRKHS::RKHS> KoopmanRKHS::RKHS_values =
-    {
-      {"Polynomial",  KoopmanRKHS::RKHS::Polynomial },
-      {"Exponential", KoopmanRKHS::RKHS::Exponential},
-      {"Gaussian",    KoopmanRKHS::RKHS::Gaussian   }
-    };
-
-
-  // RKHS kernel
-  //
-  double KoopmanRKHS::kernel(const Eigen::VectorXd& x,
-			     const Eigen::VectorXd& y)
-  {
-    if (rkhs == RKHS::Polynomial) {
-      double prod = x.adjoint()*y;
-      return pow(1.0 + prod/(d*d), alpha);
-    } else if (rkhs == RKHS::Exponential) {
-      double prod = x.adjoint()*y;
-      return exp(prod/(d*d));
-    } else if (rkhs == RKHS::Gaussian) {
-      Eigen::VectorXd diff = x - y;
-      double prod = diff.adjoint()*diff;
-      return exp(-prod/(d*d));
-    } else {
-      throw std::runtime_error("KoopmanRKHS: Unknown kernel type");
-    }
-  }
-
-  // Structure for sorting complex numbers and retrieving a permutation index
-  //
-  using complex_elem = std::pair<std::complex<double>, int>;
-  static
[…]
