### 2024-12-21 6f151e7
Draft GitHub issue template
