### 2024-12-22 ffa4513
A few more AddAccExt changes
