### 2024-12-22 28d3944
draft paper updates
