### 2024-12-22 ae2ef19
Changed 'dump' to 'dumpbasis'.  Some doxygen parameter descriptions
