### 2024-12-23 a77784e
Implementation of the pseudo-force estimator for Orient::Axis - parallels the method for computing the center motion - enhances the PseudoAccel class - adds a new member function to Component needed for Coriolis, Euler, and Centrifugal forces - makes the intent clear when we get to testing the algorithm
