### 2024-12-23 667d641
Do not apply pseudo force if EJdryrun=True, although the values will be computed and recorded in the orient log file
