### 2024-12-23 e345939
Don't call for a pseudoforce in output diagnostics if not instantiated
