### 2024-12-23 3c91cbc
Don't call for a pseudoforce in output diagnostics if not instantiated
