### 2024-12-23 5cd05f7
adding an exponential disk + Hernquist bulge model
