+++ include/DiskModels.H
@@ -223,5 +223,36 @@ public:
 
 };
 
-#endif
 
+//! The usual exponential disk + a Hernquist bulge
+class diskbulge : public EmpCylSL::AxiDisk
+{
+private:
+  double a, h, as, Mfac, d1, d2;
+
+public:
+
+  diskbulge(double a, double h, double as, double Mfac, double M=1) :
+    a(a), h(h), as(as), Mfac(Mfac), EmpCylSL::AxiDisk(M, "diskbulge")
+  {
+    params.push_back(a);
+    params.push_back(h);
+    params.push_back(as);
+    params.push_back(Mfac);
+  }
+
+  double operator()(double R, double z, double phi=0.)
+  {
+    double sh = 1.0/cosh(z/h);
+    d1 = 0.25*M*Mfac/(M_PI*a*a*h) * exp(-R/a) * sh * sh;
+
+    double rr = pow(pow(R, 2) + pow(z,2), 0.5);
+    d2 = M*(1-Mfac)*pow(as, 4)/(2.0*M_PI*rr)*pow(rr+as,-3.0);
+
+    return d1 + d2;
+
+  }
+
+};
+
+#endif
