+++ utils/ICs/cylcache.cc
@@ -165,20 +165,23 @@ void set_fpu_gdb_handler(void)
 
 // Global variables
 //
-enum DiskType { constant, gaussian, mn, exponential, doubleexpon };
+enum DiskType { constant, gaussian, mn, exponential, doubleexpon, diskbulge };
 
 std::map<std::string, DiskType> dtlookup =
   { {"constant",    DiskType::constant},
     {"gaussian",    DiskType::gaussian},
     {"mn",          DiskType::mn},
     {"exponential", DiskType::exponential},
-    {"doubleexpon", DiskType::doubleexpon}
+    {"doubleexpon", DiskType::doubleexpon},
+    {"diskbulge", DiskType::diskbulge}
   };
 
 DiskType     DTYPE;
 double       ASCALE;
 double       ASHIFT;
 double       HSCALE;
+double       HERNA;
+double       Mfac;
 double       RTRUNC  = 1.0;
 double       RWIDTH  = 0.0;
 double       ARATIO  = 1.0;
@@ -228,6 +231,18 @@ double DiskDens(double R, double z, double phi)
       ans =
 	w1*exp(-R/a1)/(4.0*M_PI*a1*a1*h1*f1*f1) +
 	w2*exp(-R/a2)/(4.0*M_PI*a2*a2*h2*f2*f2) ;
+    }
+    break;
+  case DiskType::diskbulge:
+    {
+      double f = cosh(z/HSCALE);
+      double rr = pow(pow(R, 2) + pow(z,2), 0.5);
+      double w1 = Mfac;
+      double w2 = (1-Mfac);
+      double as = HERNA;
+
+      ans = w1*exp(-R/ASCALE)/(4.0*M_PI*ASCALE*ASCALE*HSCALE*f*f) + 
+            w2*pow(as, 4)/(2.0*M_PI*rr)*pow(rr+as,-3.0) ; 
     }
     break;
   case DiskType::exponential:
@@ -327,7 +342,7 @@ main(int ac, char **av)
     ("ctype", "DiskHalo radial coordinate scaling type (one of: Linear, Log, Rat)",
      cxxopts::value<string>(ctype)->default_value("Log"))
     ("LMAXFID", "Maximum angular order for spherical basis in adaptive construction of the cylindrical basis",
-     cxxopts::value<int>(LMAXFID)->default_value("72"))
+     cxxopts::value<int>(LMAXFID)->default_value("32"))
     ("NMAXFID", "Maximum radial order for the spherical basis in adapative construction of the cylindrical basis",
      cxxopts::value<int>(NMAXFID)->default_value("32"))
     ("MMAX", "Maximum azimuthal order for the cylindrical basis",
@@ -341,7 +356,7 @@ main(int ac, char **av)
     ("NCYLODD", "Number of vertically odd basis functions per harmonic order",
      cxxopts::value<int>(NCYLODD)->default_value("6"))
     ("NMAX", "Total number of basis functions per harmonic order",
-     cxxopts::value<int>(NMAX)->default_value("18"))
+     cxxopts::value<int>(NMAX)->default_value("20"))
     ("VFLAG", "Diagnostic flag for EmpCylSL",
      cxxopts::value<int>(VFLAG)->default_value("31"))
     ("expcond", "Use analytic target density rather than particle distribution",
@@ -368,6 +383,10 @@ main(int ac, char **av)
      cxxopts::value<double>(PPower)->default_value("4.0"))
     ("DWEIGHT", "Ratio of second disk relative to the first disk for disk basis construction with double-exponential",
      cxxopts::value<double>(DWEIGHT)->default_value("1.0"))
[…]
