### 2024-12-24 4fb9894
Implementation of a Python functor for the target density function in 'pyEXP' and 'exp'
