### 2024-12-25 92c31dd
Added documenation and an example Python module
