+++ include/DiskDensityFunc.H
@@ -6,16 +6,37 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/embed.h>
 
-// Convenience
-namespace py = pybind11;
+/** Python disk-density function wrapper
 
-//! Python disk-density function wrapper
-class DiskDensityFunc
+    The wrapper class will take a user supplied Python module that
+    must supply the 'disk-density' function
+
+    An example disk-density Python module:
+    --------------------------cut here--------------------------------
+    import math
+
+    def disk_density(R, z, phi):
+        """Computes the disk density at a given radius, height, and
+        azimuth.  This would be the user's target density for basis
+        creation.
+
+        """
+        a = 1.0                         # Scale radius
+        h = 0.2                         # Scale height
+        f = math.exp(-math.fabs(z)/h)   # Prevent overflows
+        sech = 2.0*f / (1.0 + f*f)      # 
+        return math.exp(-R/a)*sech*sech/(4*math.pi*h*a*a)
+	
+    --------------------------cut here--------------------------------
+
+ */
+class __attribute__((visibility("default")))
+DiskDensityFunc
 {
 private:
 
   //! The disk-density function
-  py::function disk_density;
+  pybind11::function disk_density;
 
 public:
 
