### 2024-12-25 6c2fbcd
Duplicated Hernquist parameters for diskbulge
