### 2024-12-25 ce62bb4
Check for existence of interpreter and only start a new one if needed
