### 2024-12-25 e3e78f4
Consistency in stdout info
