+++ expui/BiorthBasis.H
@@ -302,7 +302,7 @@ namespace BasisClasses
     {
       auto [ret, worst, lworst] = orthoCompute(orthoCheck(knots));
       // For the CTest log
-      std::cout << "Spherical::orthoTest: worst=" << worst << std::endl;
+      std::cout << "---- Spherical::orthoTest: worst=" << worst << std::endl;
       return ret;
     }
 
@@ -537,7 +537,7 @@ namespace BasisClasses
     {
       auto [ret, worst, lworst] = orthoCompute(orthoCheck());
       // For the CTest log
-      std::cout << "FlatDisk::orthoTest: worst=" << worst << std::endl;
+      std::cout << "---- FlatDisk::orthoTest: worst=" << worst << std::endl;
       return ret;
     }
 
@@ -689,7 +689,7 @@ namespace BasisClasses
     {
       auto [ret, worst, lworst] = orthoCompute(sl->orthoCheck());
       // For the CTest log
-      std::cout << "Cylindrical::orthoTest: worst=" << worst << std::endl;
+      std::cout << "---- Cylindrical::orthoTest: worst=" << worst << std::endl;
       return ret;
     }
   };
@@ -833,7 +833,7 @@ namespace BasisClasses
 	}
       }
 	
-      std::cout << "Slab::orthoTest: worst=" << worst << std::endl;
+      std::cout << "---- Slab::orthoTest: worst=" << worst << std::endl;
       if (worst > __EXP__::orthoTol) return false;
       return true;
     }
@@ -955,7 +955,7 @@ namespace BasisClasses
 	}
       }
 	
-      std::cout << "Cube::orthoTest: worst=" << worst << std::endl;
+      std::cout << "---- Cube::orthoTest: worst=" << worst << std::endl;
       if (worst > __EXP__::orthoTol) return false;
       return true;
     }
