### 2024-12-25 86c8c4e
Allow the function name to be user settable, although we are not currently using this feature
