### 2024-12-26 14e5da8
Eliminate std::vector assignment warnings
