### 2024-12-26 3b84992
stdout log formating only
