### 2024-12-26 2c1b3b1
Add axis to diagnostic output.  PseudoAccel should add the original data, not the linear estimate, oops
