### 2024-12-26 dd2b6cd
Fill the state vector before approximating any values
