### 2024-12-26 11012a6
Add divide-by-zero check in quadratic regression formula
