### 2025-01-17 cf49265
Added ring generation routine, position quick start to gensph, and modernized the gen_point API
