+++ include/massmodel.H
@@ -4,6 +4,7 @@
 #include <vector>
 #include <memory>
 #include <random>
+#include <tuple>
 
 #include <Eigen/Eigen>
 
@@ -111,6 +112,10 @@ public:
 class AxiSymModel : public MassModel,
 		    public std::enable_shared_from_this<AxiSymModel>
 {
+public:
+
+  using PSret = std::tuple<Eigen::VectorXd, int>;
+
 protected:
   //@{
   //! Stuff for gen_point
@@ -126,11 +131,12 @@ protected:
   double gen_fomax, gen_ecut;
   //@}
 
-  Eigen::VectorXd gen_point_2d(int& ierr);
-  Eigen::VectorXd gen_point_2d(double r, int& ierr);
-  Eigen::VectorXd gen_point_3d(int& ierr);
-  Eigen::VectorXd gen_point_3d(double Emin, double Emax, double Kmin, double Kmax, int& ierr);
-  Eigen::VectorXd gen_point_jeans_3d(int& ierr);
+  PSret gen_point_2d();
+  PSret gen_point_2d(double r);
+  PSret gen_point_3d();
+  PSret gen_point_3d(double Emin, double Emax, double Kmin, double Kmax);
+  PSret gen_point_jeans_3d();
+  PSret gen_point_3d(double r, double theta, double phi);
   
   double Emin_grid, Emax_grid, dEgrid, dKgrid;
   vector<double> Egrid, Kgrid, EgridMass;
@@ -212,55 +218,73 @@ public:
   void set_Ecut(double cut) { gen_ecut = cut; }
 
   //! Generate a phase-space point
-  virtual Eigen::VectorXd gen_point(int& ierr) {
+  virtual PSret gen_point()
+  {
     if (dof()==2)
-      return gen_point_2d(ierr);
+      return gen_point_2d();
     else if (dof()==3)
-      return gen_point_3d(ierr);
+      return gen_point_3d();
     else
       bomb( "dof must be 2 or 3" );
     
-    return Eigen::VectorXd();
+    return {Eigen::VectorXd(), 1};
   }
   
   //! Generate a phase-space point using Jeans' equations
-  virtual Eigen::VectorXd gen_point_jeans(int& ierr) {
+  virtual PSret gen_point_jeans()
+  {
     if (dof()==2)
       bomb( "AxiSymModel: gen_point_jeans_2d(ierr) not implemented!" );
     else if (dof()==3)
-      return gen_point_jeans_3d(ierr);
+      return gen_point_jeans_3d();
     else
       bomb( "dof must be 2 or 3" );
     
-    return Eigen::VectorXd();
+    return {Eigen::VectorXd(), 1};
   }
   
   //! Generate a phase-space point at a particular radius
-  virtual Eigen::VectorXd gen_point(double r, int& ierr) {
+  virtual PSret gen_point(double r) {
     if (dof()==2)
-      return gen_point_2d(r, ierr);
+      return gen_point_2d(r);
     else if (dof()==3)
-      bomb( "AxiSymModel: gen_point(r, ierr) not implemented!" );
+      bomb( "AxiSymModel: gen_point(r) not implemented!" );
     else
       bomb( "AxiSymModel: dof must be 2 or 3" );
     
-    return Eigen::VectorXd();
+    return {Eigen::VectorXd(), 1};
+  }
+  
+  //! Generate a phase-space point at a particular radius, theta, and phi
+  virtual PSret
+  gen_point(double r, double theta, double phi)
+  {
+    if (dof()==2)
+      bomb( "AxiSymModel: gen_point(r, theta, phi) not implemented!" );
+    else if (dof()==3)
+      return gen_point_3d(r, theta, phi);
+    else
[…]
