+++ exputil/realize_model.cc
@@ -69,12 +69,10 @@ int      AxiSymModel::gen_itmax = 20000;
 const bool verbose = true;
 const double ftol = 0.01;
 
-Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
+AxiSymModel::PSret AxiSymModel::gen_point_2d()
 {
   if (!dist_defined) {
-    std::cerr << "AxiSymModel: must define distribution before realizing!"
-	      << std::endl;
-    exit (-1);
+    throw std::runtime_error("AxiSymModel: must define distribution before realizing!");
   }
 
   int it;			// Iteration counter
@@ -87,7 +85,6 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
   double Emin = get_pot(rmin);
   double Emax = get_pot(get_max_radius());
 
-
   if (gen_EJ) {
 
     if (gen_firstime) {
@@ -97,8 +94,9 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
       double dx = (Emax - Emin - 2.0*tol)/(numr-1);
       double dy = (1.0 - gen_kmin - 2.0*tol)/(numj-1);
 
+#ifdef DEBUG
       std::cout << "gen_point_2d[" << ModelID << "]: " << get_max_radius() << std::endl;
-      
+#endif
       gen_fomax = 0.0;
 
       for (int j=0; j<numr; j++) {
@@ -119,8 +117,8 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
     }
 
     
-				// Trial velocity point
-    
+    // Trial velocity point
+    //
     for (it=0; it<gen_itmax; it++) {
 
       double xxx = Emin + tol + (Emax - Emin - 2.0*tol)*Unit(random_gen);
@@ -166,7 +164,9 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
       gen_rloc.resize(gen_N);
       gen_fmax.resize(gen_N);
 
+#ifdef DEBUG
       std::cout << "gen_point_2d[" << ModelID << "]: " << get_max_radius() << std::endl;
+#endif
 
       if (rmin <= 1.0e-16) gen_logr = 0;
 
@@ -219,8 +219,8 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
     pot = get_pot(r);
     vmax = sqrt(2.0*fabs(Emax - pot));
 
-                                // Trial velocity point
-    
+    // Trial velocity point
+    //
     for (it=0; it<gen_itmax; it++) {
 
       double xxx = sqrt(Unit(random_gen));
@@ -254,13 +254,10 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
 		<< std::setw(12) << fmax
 		<< " %=" << std::setw(12)
 		<< static_cast<double>(++toomany)/totcnt << std::endl;
-    ierr = 1;
     out.setZero();
-    return out;
+    return {out, 1};
   }
 
-  ierr = 0;
-  
   cosp = cos(phi);
   sinp = sin(phi);
 
@@ -272,16 +269,14 @@ Eigen::VectorXd AxiSymModel::gen_point_2d(int& ierr)
   out[5] = vr * sinp + vt * cosp;
   out[6] = 0.0;
     
-  return out;
+  return {out, 0};
 }
 
 
-Eigen::VectorXd AxiSymModel::gen_point_2d(double r, int& ierr)
+AxiSymModel::PSret AxiSymModel::gen_point_2d(double r)
 {
   if (!dist_defined) {
[…]
