### 2025-01-17 51298ca
Update particle numbers to exactly fill each radial shell
