### 2025-01-18 7285737
Added Fibonacci spiral mapped to the sphere for uniform coverage of shells
