### 2025-01-18 846e4f4
Fixed error in velocity assignment at fixed radius
