+++ exputil/realize_model.cc
@@ -37,6 +37,7 @@
 #include <localmpi.H>
 #include <massmodel.H>
 #include <interp.H>
+#include <euler.H>
 
 #ifdef DEBUG
 #include <orbit.H>
@@ -107,9 +108,9 @@ AxiSymModel::PSret AxiSymModel::gen_point_2d()
 
 	  gen_orb.new_orbit(xxx, yyy);
 	    
-	  double zzz = distf(xxx, gen_orb.AngMom())*gen_orb.Jmax()
-	    /gen_orb.get_freq(1);
-	  gen_fomax = zzz>gen_fomax ? zzz : gen_fomax;
+	  double zzz = distf(xxx, gen_orb.AngMom())*gen_orb.Jmax()/gen_orb.get_freq(1);
+
+	  if (zzz>gen_fomax) gen_fomax = zzz;
 	}
       }
 
@@ -203,7 +204,7 @@ AxiSymModel::PSret AxiSymModel::gen_point_2d()
 	    eee = pot + 0.5*(vr*vr + vt*vt);
 
 	    double zzz = distf(eee, gen_rloc[i]*vt);
-	    fmax = zzz>fmax ? zzz : fmax;
+	    if (zzz > fmax) fmax = zzz;
 	  }
 	}
 	gen_fmax[i] = fmax*(1.0 + ftol);
@@ -232,7 +233,7 @@ AxiSymModel::PSret AxiSymModel::gen_point_2d()
 
       if (Unit(random_gen) > distf(eee, r*vt)/fmax ) continue;
 
-      if (Unit(random_gen)<0.5) vr *= -1.0;
+      if (Unit(random_gen) < 0.5) vr *= -1.0;
     
       phi = 2.0*M_PI*Unit(random_gen);
 
@@ -321,7 +322,7 @@ AxiSymModel::PSret AxiSymModel::gen_point_2d(double r)
 	  eee = pot + 0.5*(vr*vr + vt*vt);
 
 	  double zzz = distf(eee, gen_rloc[i]*vt);
-	  fmax = zzz>fmax ? zzz : fmax;
+	  if (zzz>fmax) fmax = zzz;
 	}
       }
       gen_fmax[i] = fmax*(1.0 + ftol);
@@ -347,7 +348,7 @@ AxiSymModel::PSret AxiSymModel::gen_point_2d(double r)
 
     if (Unit(random_gen) > distf(eee, r*vt)/fmax ) continue;
     
-    if (Unit(random_gen)<0.5) vr *= -1.0;
+    if (Unit(random_gen) < 0.5) vr *= -1.0;
     
     phi = 2.0*M_PI*Unit(random_gen);
 
@@ -453,7 +454,7 @@ AxiSymModel::PSret AxiSymModel::gen_point_3d()
 	  eee = pot + 0.5*(vr*vr + vt*vt);
 
 	  double zzz = distf(eee, r*vt);
-	  fmax = zzz>fmax ? zzz : fmax;
+	  if (zzz>fmax) fmax = zzz;
 	}
       }
       gen_fmax[i] = fmax*(1.0 + ftol);
@@ -625,7 +626,7 @@ AxiSymModel::PSret AxiSymModel::gen_point_3d(double r, double theta, double phi)
 	  eee = pot + 0.5*(vr*vr + vt*vt);
 
 	  double zzz = distf(eee, gen_rloc[i]*vt);
-	  fmax = zzz>fmax ? zzz : fmax;
+	  if (zzz>fmax) fmax = zzz;
 	}
       }
       gen_fmax[i] = fmax*(1.0 + ftol);
[…]
