### 2025-01-21 8539dc5
Make goldenRatio a const rather than a constexpr
