### 2025-01-23 41a1161
INSTALL.md updates [no CI]
