### 2025-01-24 4fe5c44
Print mismatched cache variables
