+++ exputil/EmpCyl2d.cc
@@ -780,19 +780,31 @@ bool EmpCyl2d::ReadH5Cache()
     auto checkInt = [&file](int value, std::string name)
     {
       int v; HighFive::Attribute vv = file.getAttribute(name); vv.read(v);
-      if (value == v) return true; return false;
+      if (value == v) return true;
+      if (myid==0) std::cout << "---- EmpCyl2d::ReadH5Cache: "
+			     << name << " expected " << value << " but found "
+			     << v << std::endl;
+      return false;
     };
 
     auto checkDbl = [&file](double value, std::string name)
     {
       double v; HighFive::Attribute vv = file.getAttribute(name); vv.read(v);
-      if (fabs(value - v) < 1.0e-16) return true; return false;
+      if (fabs(value - v) < 1.0e-16) return true;
+      if (myid==0) std::cout << "---- EmpCyl2d::ReadH5Cache: "
+			     << name << " expected " << value << " but found "
+			     << v << std::endl;
+      return false;
     };
 
     auto checkStr = [&file](std::string value, std::string name)
     {
       std::string v; HighFive::Attribute vv = file.getAttribute(name); vv.read(v);
-      if (value.compare(v)==0) return true; return false;
+      if (value.compare(v)==0) return true;
+      if (myid==0) std::cout << "---- EmpCyl2d::ReadH5Cache: "
+			     << name << " expected " << value << " but found "
+			     << v << std::endl;
+      return false;
     };
 
     //
