### 2025-02-02 #7302 @michael-petersen
@dfm , thanks for checking in! We got a little stuck working to implement one recommendation (`FetchContent`), but we think we have a path forward and will complete the response soon!
