+++ expui/expMSSA.cc
@@ -55,6 +55,12 @@
 
 namespace MSSA {
 
+  const bool useSignChoice = true;
+
+  void SvdSignChoice
+  (const Eigen::MatrixXd& X,
+   Eigen::MatrixXd& U, const Eigen::VectorXd& S, Eigen::MatrixXd& V);
+
   Eigen::MatrixXd expMSSA::wCorrKey(const Key& key)
   {
     if (RC.find(key)==RC.end()) {
@@ -289,52 +295,80 @@ namespace MSSA {
     //
     if (params["Jacobi"]) {
       // -->Using Jacobi
-      if (trajectory) {	// Trajectory matrix
+      if (trajectory) {		// Trajectory matrix
 	auto YY = Y/Scale;
 	Eigen::JacobiSVD<Eigen::MatrixXd>
 	  svd(YY, Eigen::ComputeThinU | Eigen::ComputeThinV);
 	S = svd.singularValues();
 	U = svd.matrixV();
+	if (useSignChoice) {
+	  auto V = svd.matrixU();
+	  SvdSignChoice(YY, V, S, U);
+	}
       }
       else {			// Covariance matrix
 	Eigen::JacobiSVD<Eigen::MatrixXd>
 	  svd(cov, Eigen::ComputeThinU | Eigen::ComputeThinV);
 	S = svd.singularValues();
 	U = svd.matrixU();
+	if (useSignChoice) {
+	  auto V = svd.matrixV();
+	  SvdSignChoice(cov, U, S, V);
+	}
       }
     } else if (params["BDCSVD"]) {
       // -->Using BDC
-      if (trajectory) {	// Trajectory matrix
+      if (trajectory) {		// Trajectory matrix
 	auto YY = Y/Scale;
 	Eigen::BDCSVD<Eigen::MatrixXd>
 	  svd(YY, Eigen::ComputeThinU | Eigen::ComputeThinV);
 	S = svd.singularValues();
 	U = svd.matrixV();
+	if (useSignChoice) {
+	  auto V = svd.matrixU();
+	  SvdSignChoice(YY, V, S, U);
+	}
       }
       else {			// Covariance matrix
 	Eigen::BDCSVD<Eigen::MatrixXd>
 	  svd(cov, Eigen::ComputeThinU | Eigen::ComputeThinV);
 	S = svd.singularValues();
 	U = svd.matrixU();
+	if (useSignChoice) {
+	  auto V = svd.matrixV();
+	  SvdSignChoice(cov, U, S, V);
+	}
       }
     } else {
       // -->Use Random approximation algorithm from Halko, Martinsson,
       //    and Tropp
-      if (trajectory) {	// Trajectory matrix
+      if (trajectory) {		// Trajectory matrix
 	auto YY = Y/Scale;
 	RedSVD::RedSVD<Eigen::MatrixXd> svd(YY, srank);
 	S = svd.singularValues();
 	U = svd.matrixV();
+	if (useSignChoice) {
+	  auto V = svd.matrixU();
+	  SvdSignChoice(YY, V, S, U);
+	}
       }
       else {			// Covariance matrix
 	if (params["RedSym"]) {
 	  RedSVD::RedSymEigen<Eigen::MatrixXd> eigen(cov, srank);
 	  S = eigen.eigenvalues().reverse();
 	  U = eigen.eigenvectors().rowwise().reverse();
+	  if (useSignChoice) {
+	    Eigen::MatrixXd V = U.transpose();
+	    SvdSignChoice(cov, U, S, V);
+	  }
 	} else {
 	  RedSVD::RedSVD<Eigen::MatrixXd> svd(cov, srank);
 	  S = svd.singularValues();
[…]
