### 2025-02-13 46161f9
Fix typo in dimension check
