### 2025-02-13 8661d07
Updates to improve speed; toggle sign choice
