### 2025-02-14 f19fcd4
Turn on unique sign algorithm by default
