### 2025-02-14 55a6190
Use OpenMP in sign choice computation
