### 2025-02-20 72fc23b
Removed the angle-only quick start for the sphere, which leaves too much even-order noise to be worth keeping
