+++ expui/expMSSA.H
@@ -249,7 +249,15 @@ namespace MSSA
        @param toTerm write to stdout if true
        @param toFile write to file if true
     */
-    void kmeans(int clusters, bool toTerm=true, bool toFile=false);
+    void kmeansPrint(int clusters, bool toTerm=true, bool toFile=false);
+
+    //! Get Kmeans analysis per channel
+    std::tuple<std::vector<int>, std::vector<double>, double>
+    kmeansChannel(int clusters, Key key);
+
+    //! Get Kmeans analysis for all channels
+    std::tuple<std::vector<int>, std::vector<double>, double>
+    kmeans(int clusters);
 
     //! Save current MSSA state to an HDF5 file with the given prefix
     void saveState(const std::string& prefix);
