### 2025-02-25 4f40e7c
Allow the user to specify the seeding strategy for k-means.  2-strided by default, a good choice for quasi-periodic systems
