### 2025-02-27 888632a
Overload CoefStruct::setCoefs() to have two call signatures
