### 2025-03-12 EXP-code/EXP#95 @michael-petersen
`EXP-docs` piece is being addressed by https://github.com/EXP-code/EXP-docs/pull/32.

I believe that the `pyEXP-examples` `README.md` issue has been addressed already?
