+++ CMakeLists.txt
@@ -68,7 +68,7 @@ find_package(OpenMP)
 find_package(FFTW REQUIRED)
 find_package(HDF5 COMPONENTS C CXX HL REQUIRED)
 find_package(TIRPC)	       # Check for alternative Sun rpc support
-find_package(Eigen3 3.4 REQUIRED)
+find_package(Eigen3 3.4...<3.5 REQUIRED)
 find_package(PNG)
 
 # Check for FE
