### 2025-03-12 EXP-code/EXP#94 @michael-petersen
We've elected to go for the 'specify requirements more clearly' route for now -- see https://github.com/EXP-code/EXP-docs/pull/32.

Thanks for the pointers to `FetchContent`! We're still working on that solution across platforms, hopefully for a later version.
