### 2025-03-12 EXP-code/EXP#96 @michael-petersen
We've sorted this by requiring Eigen 3.4 and above -- see https://github.com/EXP-code/EXP-docs/pull/32.
