### 2025-03-12 0ff5000
add CITATIONS.bib
