+++ CONTRIBUTING.md
@@ -2,7 +2,8 @@
 
 There are many ways to contribute to EXP. Here are some of them:
 
--   Blog about EXP. Cite the EXP published papers. Tell the world how
+-   Blog about EXP. Cite the EXP published papers (using 
+    the papers in `CITATIONS.bib`). Tell the world how
     you're using EXP. This will help newcomers with more examples and
     will help the EXP project to increase its visibility.
 -   Report bugs and request features in the [issue
