### 2025-03-12 EXP-code/EXP#95 @The9Cat
The pyEXP-examples Recipes README.md is consistent; the directories all exist.  Some of these directories are placeholders, however, waiting for contributions.
