### 2025-03-12 EXP-code/EXP#95 @michael-petersen
The `README.md` links have been sorted out in [cc29e51](https://github.com/EXP-code/pyEXP-examples/pull/13/commits/cc29e512c21496ea7b5950cecdbef296daf9731f), along with an explanation behind the structure.
