### 2025-03-12 #7302 @michael-petersen
Hi @pmocz , @schuhmaj , we've finally sorted a bunch of last little issues! Everything is logged in https://github.com/EXP-code/EXP/pull/100, and hopefully links out to the all the comments (some in this thread, some as issues). Could you take a look and let us know if this addresses all the points? Thank you!
