### 2025-03-15 4ab47e9
Added logic to cache and fix the monopole coefficients
