### 2025-03-25 cd8cbde
Add KDE 2d smoothing to diffpsp
