+++ utils/PhaseSpace/KDE2d.cc
@@ -1,0 +1,262 @@
+#include <KDE2d.H>
+
+namespace KDE
+{
+  // Function to compute the Gaussian kernel in pixel units
+  //
+  std::vector<double> KDE2d::gaussian_kernel_2d(int xsize, int ysize)
+  {
+    std::vector<double> kernel(xsize * ysize);
+    double sum = 0.0;
+    for (int i = -xsize / 2; i < xsize / 2 + 1; i++) {
+
+      double dx = delx*i/numx;
+      double facx = exp(-0.5*dx*dx/(sigmax*sigmax));
+  
+      for (int j = -ysize / 2; j < ysize / 2 + 1; j++) {
+
+	double dy = dely*j/numy;
+	double facy = exp(-0.5*dy*dy/(sigmay*sigmay)); 
+
+	kernel[(i + xsize / 2) * ysize + (j + ysize / 2)] = facx * facy;
+	
+	sum += facx * facy;
+      }
+    }
+    
+    // Normalize the kernel
+    //
+    for (auto & v : kernel) v /= sum;
+
+    return kernel;
+  }
+
+  
+  void KDE2d::kde_fft_2d()
+  {
+    // Sanity check
+    //
+    if (numx != grid.rows() || numy != grid.cols())
+      throw std::runtime_error("Invalid grid dimensions");
+
+    // Compute the Gaussian kernel
+    //
+    int kern_xsize = std::max<int>(minKsize,
+				   std::floor(sigmax/delx*numx*minKsize));
+
+    if (kern_xsize % 2 == 0) kern_xsize += 1; // Make the size odd
+    
+    int kern_ysize = std::max<int>(minKsize,
+				   std::floor(sigmay/dely*numy*minKsize));
+
+    if (kern_ysize % 2 == 0) kern_ysize += 1; // Make the size odd
+    
+    std::vector<double> kern = gaussian_kernel_2d(kern_xsize, kern_ysize);
+    
+    if (debug) {
+      std::ofstream fout("kern0.dat");
+      fout << kern_xsize << " " << kern_ysize << std::endl;
+      for (int j=0; j<kern_ysize; j++) {
+	for (int i=0; i<kern_xsize; i++) {
+	  fout << std::setw( 8) << i 
+	       << std::setw( 8) << j
+	       << std::setw(18) << kern[i*kern_ysize + j] << std::endl;
+	}
+      }
+    }
+    
+    // Pad the grid and kernel to the same size
+    //
+    int padded_xsize = numx + kern_xsize - 1;
+    int padded_ysize = numy + kern_ysize - 1;
+    
+    std::vector<double> padded_grid(padded_xsize * padded_ysize, 0.0);
+    std::vector<double> padded_kern(padded_xsize * padded_ysize, 0.0);
+    
+    for (int i=0; i<numx; i++) {
+      for (int j=0; j<numy; j++) {
+	padded_grid[i*padded_ysize + j] = grid(i, j);
+      }
+    }
+    
+
+    for (int i=0; i<kern_xsize; i++) {
+
+      int ii = i - kern_xsize/2;
+      if (ii < 0) ii += padded_xsize;
+      
+      for (int j=0; j<kern_ysize; j++) {
+	
+	int jj = j - kern_ysize/2;
+	if (jj < 0) jj += padded_ysize;
+	
+	padded_kern[ii*padded_ysize + jj] = kern[i*kern_ysize + j];
+      }
+    }
+
+    if (debug) {
+      std::ofstream kdat("kern.dat");
+      std::ofstream ddat("data.dat");
+      std::ofstream gdat("grid.dat");
+      kdat << padded_xsize << " " << padded_ysize << std::endl;
[…]
