### 2025-03-27 b5d7ced
Expose the field basis label list for pybind11; this adds previously missing functionality
