### 2025-03-27 d7c0e7e
Typo in spherical velocity conversion
