### 2025-04-01 85b649b
Added orthoTest for input SLGridSph basis input to EmpCylSL
