+++ exputil/EmpCylSL.cc
@@ -16,6 +16,7 @@
 #include <thread>
 #include <exp_thread.h>
 #include <EXPException.H>
+#include <exputils.H>
 
 #include <Eigen/Eigenvalues>
 
@@ -391,7 +392,7 @@ void EmpCylSL::reset(int numr, int lmax, int mmax, int nord,
   ortho = std::make_shared<SLGridSph>(make_sl(), LMAX, NMAX, NUMR,
 				      RMIN, RMAX*0.99, false, 1, 1.0);
 
-  orthoTest(ortho->orthoCheck(std::max<int>(nmax*50, 200)), "EmpCylSL", "l");
+  orthoTest(ortho->orthoCheck(std::max<int>(NMAX*50, 200)), "EmpCylSL[SLGridSph]", "l");
 
   // Resize (should not be necessary) but just in case some future
   // feature changes mulitstep on the fly
@@ -870,7 +871,7 @@ int EmpCylSL::read_eof_file(const string& eof_file)
   ortho = std::make_shared<SLGridSph>(make_sl(), LMAX, NMAX, NUMR,
 					RMIN, RMAX*0.99, false, 1, 1.0);
 
-  orthoTest(ortho->orthoCheck(std::max<int>(nmax*50, 200)), "EmpCylSL", "l");
+  orthoTest(ortho->orthoCheck(std::max<int>(NMAX*50, 200)), "EmpCylSL[SLGridSph]", "l");
 
   setup_eof();
   setup_accumulation();
@@ -1442,7 +1443,7 @@ void EmpCylSL::compute_eof_grid(int request_id, int m)
     ortho = std::make_shared<SLGridSph>(make_sl(), LMAX, NMAX, NUMR,
 					  RMIN, RMAX*0.99, false, 1, 1.0);
 
-    orthoTest(ortho->orthoCheck(std::max<int>(nmax*50, 200)), "EmpCylSL", "l");
+    orthoTest(ortho->orthoCheck(std::max<int>(NMAX*50, 200)), "EmpCylSL[SLGridSph]", "l");
   }
 
 
@@ -1625,7 +1626,7 @@ void EmpCylSL::compute_even_odd(int request_id, int m)
 					  LMAX, NMAX, NUMR, RMIN, RMAX*0.99,
 					  false, 1, 1.0);
 
-    orthoTest(ortho->orthoCheck(std::max<int>(nmax*50, 200)), "EmpCylSL", "l");
+    orthoTest(ortho->orthoCheck(std::max<int>(NMAX*50, 200)), "EmpCylSL[SLGridSph]", "l");
   }
 
   double dens, potl, potr, pott;
@@ -2307,7 +2308,7 @@ void EmpCylSL::generate_eof(int numr, int nump, int numt,
     ortho = std::make_shared<SLGridSph>(make_sl(), LMAX, NMAX, NUMR,
 					  RMIN, RMAX*0.99, false, 1, 1.0);
 
-    orthoTest(ortho->orthoCheck(std::max<int>(nmax*50, 200)), "EmpCylSL", "l");
+    orthoTest(ortho->orthoCheck(std::max<int>(NMAX*50, 200)), "EmpCylSL[SLGridSph]", "l");
   }
 
 
@@ -2604,7 +2605,7 @@ void EmpCylSL::accumulate_eof(double r, double z, double phi, double mass,
     ortho = std::make_shared<SLGridSph>
       (make_sl(), LMAX, NMAX, NUMR, RMIN, RMAX*0.99, false, 1, 1.0);
 
-    orthoTest(ortho->orthoCheck(std::max<int>(nmax*50, 200)), "EmpCylSL", "l");
[…]
