### 2025-04-01 ac5a8e3
Fix a few missing defaults
