### 2025-04-01 ae75211
Preliminary Cuda implementation for UserBar
