+++ src/user/cudaUserBar.cu
@@ -1,0 +1,192 @@
+// -*- C++ -*-
+
+#include <cudaUtil.cuH>
+#include <cudaReduce.cuH>
+
+#include <Component.H>
+#include "UserSat.H"
+
+// Global device symbols for CUDA kernel
+//
+__device__ __constant__
+cuFP_t userBarAmp, userBarPosAng, userBarCen[3], userBarB5;
+
+__device__ __constant__
+int userBarSoft;
+
+
+// Cuda implementation of bar force
+//
+__global__ void
+userBarForceKernel(dArray<cudaParticle> P, dArray<int> I,
+		   int stride, PII lohi)
+{
+  const int tid = blockDim.x * blockIdx.x + threadIdx.x;
+
+  for (int n=0; n<stride; n++) {
+    int i     = tid*stride + n;	// Index in the stride
+    int npart = i + lohi.first;	// Particle index
+
+    // Temporary relative position
+    cuFP_t fpos[3];
+
+    // Angular dependence from position angle
+    cuFP_t cos2p = cos(2.0*userBarPosAng);
+    cuFP_t sin2p = sin(2.0*userBarPosAng);
+
+    if (npart < lohi.second) {
+      
+#ifdef BOUNDS_CHECK
+      if (npart>=P._s) printf("out of bounds: %s:%d\n", __FILE__, __LINE__);
+#endif
+      cudaParticle & p = P._v[I._v[npart]];
+      
+      // Compute position relative to bar
+      cuFP_t rr = 0.0, fac, ffac, nn;
+      for (int k=0; k<3; k++) {
+	fpos[k] = p.pos[k] - userBarCen[k];
+	rr += fpos[k] * fpos[k];
+      }
+      rr = sqrt(rr);
+
+      cuFP_t xx = fpos[0];
+      cuFP_t yy = fpos[1];
+      cuFP_t zz = fpos[2];
+      cuFP_t pp = (xx*xx - yy*yy)*cos2p + 2.0*xx*yy*sin2p;
+    
+      if (userBarSoft) {
+	fac = 1.0 + rr/b5;
+	ffac = -userBarAmp / pow(fac, 6.0);
+	nn = pp / (b5*rr);
+      } else {
+	fac = 1.0 + pow(rr/b5, 5.0);
+	ffac = -userBarAmp / (fac*fac);
+	nn = pp * pow(rr/b5, 3.0)/ (b5*b5);
+      }
+
+      // Add acceleration
+      p.acc[0] += ffac * ( 2.0*( xx*cos2p + yy*sin2p)*fac - 5.0*nn*xx);
+      p.acc[1] += ffac * ( 2.0*(-yy*cos2p + xx*sin2p)*fac - 5.0*nn*yy);
+      p.acc[2] += ffac * ( -5.0*nn*zz);
+
+      // Add external potential
+      p.potext += -ffac*pp*fac;
+      
+    } // Particle index block
+
+  } // END: stride loop
+
+}
+
+
+__global__
+void testConstantsUserSat(cuFP_t tnow)
+{
+  printf("-------------------------\n");
+  printf("---UserSat constants-----\n");
+  printf("-------------------------\n");
+  printf("   Time   = %e\n", tnow          );
+  printf("   Amp    = %e\n", userBarAmp    );
+  printf("   Amp    = %e\n", userBarNumFac );
+  printf("   b5     = %e\n", userBarB5     );
+  printf("   Center = %e, %e, %e\n",
+	 userBarCen[0], userBarCen[1], userBarCen[2] );
+  if (userBarSoft) printf("   Soft   = true\n" );
+  else             printf("   Soft   = false\n");
[…]
