+++ src/user/UserBar.H
@@ -31,6 +31,11 @@ private:
   void * determine_acceleration_and_potential_thread(void * arg);
   void initialize();
 
+#if HAVE_LIBCUDA==1
+  //! Cuda implementation
+  void determine_acceleration_and_potential_cuda();
+#endif
+
   double length, bratio, cratio, amplitude, Ton, Toff, DeltaT, Fcorot;
   bool fixed, soft;
   string filename;
