### 2025-04-01 ab77174
Simply CUDA stanza in CMakeLists.txt
