### 2025-04-01 afd3328
Fix up a bunch of typos in UserBar cuda implementation
