### 2025-04-01 b870b3b
Reduce rank of correlation matrix to match reconstruction dimension
