### 2025-04-03 #7302 @schuhmaj
@dfm I finished my review and I fully recommend the acceptance of `EXP` into JOSS.

@michael-petersen @The9Cat Thank you very much for considering my feedback and providing this amazing software to the Open-Source Community 😀
