### 2025-04-06 #7302 @michael-petersen
@dfm, thanks! The edits look great. I've gone back through the manuscript and I'm happy.

The new version number is 7.8.0.

The 7.8.0 version of the software is now archived on Zenodo, with DOI 10.5281/zenodo.15164151.

@schuhmaj and @pmocz, thank you so much for your reviews.

@The9Cat, we made it!
