### 2025-04-07 6cfc33e
Bump devel version to match main v7.8.0 release
