### 2025-04-08 962c3c9
Added doc for new 'wCorr*' behavior
