+++ pyEXP/MSSAWrappers.cc
@@ -366,6 +366,12 @@ void MSSAtoolkitClasses(py::module &m) {
 
         Notes
         -----
+        The w-correlation matrix needs the reconstructed trajectory matrices for each
+        of the eigenvalue, PC pairs.  Calling this method will recompute the reconstruction
+        for all eigenvalues up to 'nPC' and return an nPC x nPC matrix.  If the 'nPC'
+        parameter is not specified, it will be set to `numpc` used to construct the 
+        instance.  Any previous reconstruction will be overwritten.
+
         Returns the combined cosine+sine correlation for complex types for viewing 
         (e.g., with 'imshow').
 
@@ -390,8 +396,15 @@ void MSSAtoolkitClasses(py::module &m) {
 
         Notes
         -----
-        The index key here is 'extended' by the prefixed component index
+        The index key here is 'extended' by the prefixed component index.
   
+        Computation of the w-correlation matrix needs the reconstructed
+        trajectory matrices for each of the (eigenvalue, PC) pairs.  Calling
+        this method will recompute the reconstruction for all eigenvalues up to
+        order 'npc' and return an (nPC x nPC) matrix.  If the 'nPC' parameter is
+        not specified, it will be set to the `numpc` used in the original
+        construction.  Any prior reconstruction will be overwritten.
+
         The rows and columns contain distinct cosine and sine indicies if the channel 
         is complex valued.
 
@@ -419,8 +432,15 @@ void MSSAtoolkitClasses(py::module &m) {
 
         Notes
         -----
-        The w-correlation values range from 0 to 1, where a higher value corresponds to a 
-        stronger correlation.
+        The w-correlation values range from 0 to 1, where a higher value
+        corresponds to a stronger correlation.
+
+        Computation of the w-correlation matrix needs the reconstructed
+        trajectory matrices for each of the (eigenvalue, PC) pairs.  Calling
+        this method will recompute the reconstruction for all eigenvalues up to
+        order 'npc' and return an (nPC x nPC) matrix.  If the 'nPC' parameter is
+        not specified, it will be set to the `numpc` used in the original
+        construction.  Any prior reconstruction will be overwritten.
 
         See also
         --------
@@ -441,8 +461,15 @@ void MSSAtoolkitClasses(py::module &m) {
 
         Notes
         -----
-        The w-correlation values range from 0 to 1, where a higher value corresponds to a 
-        stronger correlation.
+        The w-correlation values range from 0 to 1, where a higher value
+        corresponds to a stronger correlation.
+
+        Computation of the w-correlation matrix needs the reconstructed
+        trajectory matrices for each of the (eigenvalue, PC) pairs.  Calling
+        this method will recompute the reconstruction for all eigenvalues up to
+        order 'npc' and return an (nPC x nPC) matrix.  If the 'nPC' parameter is
+        not specified, it will be set to the `numpc` used in the original
+        construction.  Any prior reconstruction will be overwritten.
 
         See also
         --------
@@ -455,17 +482,18 @@ void MSSAtoolkitClasses(py::module &m) {
 	py::arg("clusters") = 4,
 	py::arg("stride") = 2,
 	R"(
-        Do a k-means analysis on the reconstructed trajectory matrices for a single channel (specified key value) to
-        provide grouping insight.  A vector of channel indices that identify clusters is return in a vector ordered by PC
-        index.
+        Do a k-means analysis on the reconstructed trajectory matrices for a
+        single channel (specified key value) to provide grouping insight.  A
+        vector of channel indices that identify clusters is return in a vector
+        ordered by PC index.
 
         Parameters
         ----------
         clusters : int, default=4
             number of clusters for the k-means analysis
         stride : int, default=2
-            if positive, the initial cluster centers are stride selected from the PC list.  If zero, the centers are
[…]
