### 2025-04-10 fe274b7
Final checked version
