### 2025-04-10 c96ebc5
Some very minor clean up of warning output
