### 2025-04-12 #7302 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
