### 2025-04-15 149a493
Expose coefctr from Basis to client classes
