+++ expui/sgsmooth.cc
@@ -1,0 +1,587 @@
+//! 
+// Sliding window signal processing (and linear algebra toolkit).
+//
+// supported operations:
+// <ul>
+// <li> Savitzky-Golay smoothing.
+// <li> computing a numerical derivative based of Savitzky-Golay smoothing.
+// <li> required linear algebra support for SG smoothing using STL based
+//      vector/matrix classes 
+// </ul>
+//
+// \brief Linear Algebra "Toolkit".
+//
+
+// system headers
+#include <stdio.h>
+#include <stddef.h>             // for size_t
+#include <math.h>               // for fabs  
+#include <vector>
+
+#include "sgsmooth.h"
+
+//! default convergence
+static const double TINY_FLOAT = 1.0e-300;
+
+//! comfortable array of doubles
+typedef std::vector<double> float_vect;
+//! comfortable array of ints;
+typedef std::vector<int>    int_vect;
+
+// savitzky golay smoothing.
+static float_vect sg_smooth(const float_vect &v, const int w, const int deg);
+//! numerical derivative based on savitzky golay smoothing.
+static float_vect sg_derivative(const float_vect &v, const int w, 
+                                const int deg, const double h=1.0);
+
+// c-callable interface to for tcl plugin code
+extern "C" double *calc_sgsmooth(const int ndat, double *input,
+                      const int window, const int order)
+{
+    float_vect in(ndat);
+    int i;
+
+#if defined(_OPENMP)
+#pragma omp parallel for private(i) schedule(static)
+#endif    
+    for (i=0; i < ndat; ++i) {
+        in[i] = input[i];
+    }
+
+    float_vect out = sg_smooth(in, window, order);
+
+#if defined(_OPENMP)
+#pragma omp parallel for private(i) schedule(static)
+#endif    
+    for (i=0; i < ndat; ++i) {
+        input[i] = out[i];
+    }
+
+    return input;
+}
+    
+
+extern "C" double *calc_sgsderiv(const int ndat, double *input,
+                      const int window, const int order, 
+                      const double delta)
+{
+    float_vect in(ndat);
+    int i;
+
+#if defined(_OPENMP)
+#pragma omp parallel for private(i) schedule(static)
+#endif    
+    for (i=0; i < ndat; ++i) {
+        in[i] = input[i];
+    }
+
+    float_vect out = sg_derivative(in, window, order, delta);
+
+#if defined(_OPENMP)
+#pragma omp parallel for private(i) schedule(static)
+#endif    
+    for (i=0; i < ndat; ++i) {
+        input[i] = out[i];
+    }
+
+    return input;
+}
+
+/*! matrix class.
+ *
+ * This is a matrix class derived from a vector of float_vects.  Note that
+ * the matrix elements indexed [row][column] with indices starting at 0 (c
+ * style). Also note that because of its design looping through rows should
+ * be faster than looping through columns.  
+ *
+ * \brief two dimensional floating point array
+ */
+class float_mat : public std::vector<float_vect> {
+private:
+    //! disable the default constructor
+    explicit float_mat() {};
+    //! disable assignment operator until it is implemented.
+    float_mat &operator =(const float_mat &) { return *this; };
+public:
+    //! constructor with sizes
+    float_mat(const size_t rows, const size_t cols, const double def=0.0);
+    //! copy constructor for matrix
+    float_mat(const float_mat &m);
+    //! copy constructor for vector
+    float_mat(const float_vect &v);
+
[…]

[…]
