### 2025-04-15 970474f
Expose coefctr from Basis to client classes
