### 2025-04-15 ec05fd9
Add omitted center subtraction for getFields evaluation in evalaccel()
