+++ expui/BiorthBasis.cc
@@ -3905,14 +3905,10 @@ namespace BasisClasses
     //
     int numT = floor( (tfinal - tinit)/h );
 
-    // Number of output steps.  Will attempt to find the best stride...
+    // Compute output step
     //
-    int stride = std::ceil(static_cast<double>(numT)/static_cast<double>(nout));
-    if (stride>1) numT = nout * stride;
-
-    // Compute the interval-matching step
-    //
-    h = (tfinal - tinit)/(numT - 1);
+    nout = std::min<int>(numT, nout);
+    double H = (tfinal - tinit)/nout;
 
     // Return data
     //
@@ -3944,11 +3940,9 @@ namespace BasisClasses
       for (int k=0; k<6; k++) ret(n, k, 0) = ps(n, k);
 
     double tnow = tinit;
-    int s = 0, cnt = 0;
-    while (s < numT) {
-      if (tfinal - tnow < h) h = tfinal - tnow;
+    for (int s=1, cnt=1; s<numT; s++) {
       std::tie(tnow, ps) = OneStep(tnow, h, ps, accel, bfe, F);
-      if (s++ % stride == 0) {
+      if (tnow >= H*cnt-h*1.0e-8) {
 	times(cnt) = tnow;
 	for (int n=0; n<rows; n++)
 	  for (int k=0; k<6; k++) ret(n, k, cnt) = ps(n, k);
