### 2025-04-17 804d665
Fix round off boundary and never ignore nout
