### 2025-04-21 9fe6efa
add readonly attributes
