### 2025-04-22 dd1105c
Fixed the flipped logic of inertial vs non-inertial center subtraction
