### 2025-04-22 8accceb
Restore Naccel setting, oops
