### 2025-04-23 b010474
Put the evaluation point at the center of the interval when possible
