### 2025-04-23 47edaf1
Added a inertial coodindate reset member; fixed an interpolation fence-post error that did not affect the results.
