### 2025-04-23 f792b9c
Add a bit of tolerance to the center array grid interpolation before flagging out of bounds
