+++ expui/BasisFactory.cc
@@ -359,18 +359,30 @@ namespace BasisClasses
   {
     Eigen::Vector3d ret;
 
-    if (time < t_accel(0) || time > t_accel(t_accel.size()-1)) {
-      std::cout << "ERROR: " << time << " is outside of range of the non-inertial DB"
-		<< std::endl;
-      ret.setZero();
-      return ret;
+    auto n = t_accel.size();
+
+    // Allow a little bit of buffer in the allowable on-grid range but
+    // otherwise force termination
+    //
+    if ( time < t_accel(0  ) - 0.5*(t_accel(1  ) - t_accel(0  ))  ||
+	 time > t_accel(n-1) + 0.5*(t_accel(n-1) - t_accel(n-2)) ) {
+      
+      std::ostringstream msg;
+      msg << "Basis::currentAccel: " << time
+	  << " is outside the range of the non-inertial DB ["
+	  << t_accel(0) << ", " << t_accel(n-1) << "]";
+
+      throw std::runtime_error(msg.str());
     }
+    // Do the quadratic interpolation
+    //
     else {
       int imin = 0;
-      int imax = std::lower_bound(t_accel.data(), t_accel.data()+t_accel.size(), time) - t_accel.data();
+      int imax = std::lower_bound(t_accel.data(), t_accel.data()+n, time) - t_accel.data();
 
       // Get a range around the current time of approx size Naccel
-      imax = std::min<int>(t_accel.size()-1, imax + Naccel/2);
+      //
+      imax = std::min<int>(n-1, imax + Naccel/2);
       imin = std::max<int>(imax - Naccel, 0);
 
       int num = imax - imin + 1;
