### 2025-04-23 c3a2275
Assign passed coefficients as the coefret internal basis instance so this can be queried by clients; added pseudo accel debug output for testing
