### 2025-04-24 c74d5b3
Some cleanup; turn of RK4 test and pseudo force test output
