### 2025-04-24 e4a977c
Apply suggestions from code review
Typos in docstrings only.

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
