### 2025-04-28 #7302 @warrickball
Hi @michael-petersen! I'm the relevant associate editor-in-chief and will take this through the publication process. I've been away for most of April but am working through my JOSS backlog and hope to at least start on processing this in the next few working days. Sorry for the extra wait!
