### 2025-04-29 2288967
Adjust function repack data read as row-major in col-major order to deal with HighFive row-major storage
