### 2025-04-29 40d22f7
Remove unused 'NewCache' boolean; we might also consider removing 'NewCoefs' and completely deprecreate the old binary cache format
