### 2025-04-29 da60a3b
Added deprecation warntings; tested with the two cache files I used previously
