### 2025-04-29 4de3767
add specific deprecation warning for 7.9
