### 2025-05-02 #7302 @warrickball
I've done some preliminary checks of the submission and hope to check the paper over the next few days (after which I'll open a PR with editorial changes). In the meantime, note that the Zenodo archive metadata says this is version v1, rather than 7.8.0. I presume 7.8.0 is correct, as indicated in the actual filename of the tarball in the archive. You should be able to edit the Zenodo record's metadata to fix this, and *not* have to create a new Zenodo archive.
