### 2025-05-03 #7302 @michael-petersen
@warrickball, thank you for catching this! I've updated the Zenodo archive metadata.
