### 2025-05-03 #7302 @warrickball
Great! I'm having an editorial look at the paper now and noticed that the Zenodo metadata says the licence is CC-BY-4.0 but it looks like it should be GPL-3.0 (or something similar). Could you change that too? Thanks!
