### 2025-05-04 be69854
Initial editorial tweaks for JOSS paper
