+++ Paper/paper/paper.md
@@ -18,7 +18,7 @@ authors:
 affiliations:
  - name: University of Edinburgh, UK
    index: 1
- - name: University of Massachusetts/Amherst, USA
+ - name: University of Massachusetts Amherst, USA
    index: 2
 date: 01 June 2024
 bibliography: paper.bib
@@ -39,16 +39,16 @@ bodies to evolve under their mutual gravity are capable of reproducing
 such complexities but robust links to fundamental theoretical
 explanations are still missing.
 
-Basis Function Expansions (BFE) represent fields as a linear
+Basis Function Expansions (BFEs) represent fields as a linear
 combination of orthogonal functions. BFEs are particularly well-suited
 for studies of perturbations from equilibrium, such as the evolution
 of a galaxy.  For any galaxy simulation, a biorthogonal BFE can fully
 represent the density, potential and forces by time series of
 coefficients.  The coefficients have physical meaning: they represent
-the gravitational potential energy in a given function.  The variation
+the gravitational potential energy in a given function.  The variation of
 the function coefficients in time encodes the dynamical evolution.
-The representation of simulation data by BFE results in huge
-compression of the information in the dynamical fields; for example,
+The representation of simulation data by BFEs results in huge
+compression of the information in the dynamical fields.  For example,
 1.5 TB of phase space data enumerating the positions and velocities of
 millions of particles becomes 200 MB of coefficient data!
 
@@ -73,7 +73,7 @@ The need for methodology that seamlessly connects theoretical
 descriptions of dynamics, N-body simulations, and compact descriptions
 of observed data gave rise to `EXP`. This package provides recent
 developments from applied mathematics and numerical computation to
-represent complete series of _Basis Function Expansions_ that describe
+represent complete series of BFEs that describe
 the variation of _any_ field in space.  In the context of galactic
 dynamics, these fields may be density, potential, force, velocity
 fields or any intrinsic field produced by simulations such as
@@ -85,11 +85,11 @@ simulated, and by analogy, observed data.
 
 `EXP` decomposes a galaxy into multiple bases for a variety of scales
 and geometries and is thus able to represent arbitrarily complex
-simulation with many components (e.g., disk, bulge, dark matter halo,
+simulations with many components (e.g., disk, bulge, dark matter halo,
 satellites).  `EXP` is able to efficiently summarize the degree and
 nature of asymmetries through coefficient amplitudes tracked through
 time and provide details at multiple scales. The
-amplitudes themselves enable ex-post-facto dynamical discovery.  
+amplitudes themselves enable ex-post-facto dynamical discovery.
 `EXP` is a collection of object-oriented C++ libraries with an
 associated modular N-body code and a suite of stand-alone analysis
 applications.
@@ -106,11 +106,11 @@ coefficients and discover dynamical relationships using time series
 analysis such as mSSA.  We provide a [full online
 manual](https://exp-docs.readthedocs.io) hosted by ReadTheDocs.
 
-The software package brings published -- but difficult to implement --
-applied-math technologies into the astronomical mainstream.  `EXP` and
+The software package brings published---but difficult to implement---applied-math
+technologies into the astronomical mainstream.  `EXP` and
 the associated Python interface `pyEXP` accomplish this by providing
 tools integrated with the Python ecosystem, and in particular are
-well-suited for interactive Python [@iPython] use through (e.g.)
+well-suited for interactive Python [@iPython] use through, e.g.,
 Jupyter notebooks [@jupyter]. `EXP` serves as the
 scaffolding for new imaginative applications in galactic dynamics,
 providing a common dynamical language for simulations and analytic
@@ -126,7 +126,7 @@ computed bases and resulting coefficient data are stored in HDF5
 
 | Name        | Description |
[…]
