### 2025-05-04 #7302 @warrickball
I've just opened a [pull request](https://github.com/EXP-code/EXP/pull/138) with an initial set of editorial tweaks. Once those changes (or an agreed subset) are merged I'll do another test run of the publication pipeline.
