### 2025-05-05 #7302 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
