### 2025-05-05 #7302 @warrickball
Congratulations @michael-petersen & @The9Cat, your paper has been published in JOSS!
